American River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

american_river_floodplain

Format

dataframe with 35 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run and Steelhead Modeling

The entire mapped rearing extent of 22.81 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Active channel area of 670.20 acres estimated through remote sensing analysis was subtracted from total inundated area to get inundated floodplain area.

Examples

american_river_floodplain

#> # A tibble: 35 x 5
#>    flow_cfs FR_floodplain_acr… SR_floodplain_acr… ST_floodplain_ac… watershed  
#>       <dbl>              <dbl>              <dbl>             <dbl> <chr>      
#>  1      200               0                  0                 0    American R…
#>  2     1200               0                  0                 0    American R…
#>  3     2200               6.00               6.00              6.00 American R…
#>  4     3200             160.               160.              160.   American R…
#>  5     4200             228.               228.              228.   American R…
#>  6     5200             269.               269.              269.   American R…
#>  7     5570             294.               294.              294.   American R…
#>  8     6200             338.               338.              338.   American R…
#>  9     7200             390.               390.              390.   American R…
#> 10     8200             455.               455.              455.   American R…
#> # ... with 25 more rows