Antelope Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

antelope_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. An Antelope scaling factor of 47% was calculated as the average ratio of Antelope Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 10.5 mile low gradient and 21.1 mile high gradient mapped rearing extent in Antelope Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 199 acres) and floodplain inundation area.

Examples

antelope_creek_floodplain

#>      flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1    32.69259            0.000000            0.000000            0.000000
#> 2    79.39630            0.000000            0.000000            0.000000
#> 3   126.10000            0.000000            0.000000            0.000000
#> 4   172.80371            0.000000            0.000000            0.000000
#> 5   219.50742            0.000000            0.000000            0.000000
#> 6   266.21112            0.000000            0.000000            0.000000
#> 7   266.67816            0.000000            0.000000            0.000000
#> 8   312.91483            0.000000            0.000000            0.000000
#> 9   359.61853            0.000000            0.000000            0.000000
#> 10  406.32224            1.230719            1.230719            1.230719
#> 11  453.02594            3.614342            3.614342            3.614342
#> 12  499.72965            5.613619            5.613619            5.613619
#> 13  546.43335            7.763825            7.763825            7.763825
#> 14  593.13706            9.664965            9.664965            9.664965
#> 15  639.84076           11.692131           11.692131           11.692131
#> 16  686.54447           13.605855           13.605855           13.605855
#> 17  733.24818           15.457140           15.457140           15.457140
#> 18  779.95188           17.046280           17.046280           17.046280
#> 19  826.65559           18.567412           18.567412           18.567412
#> 20  873.35929           19.998262           19.998262           19.998262
#> 21  920.06300           21.261382           21.261382           21.261382
#> 22  966.76670           22.591707           22.591707           22.591707
#> 23 1013.47041           23.628220           23.628220           23.628220
#> 24 1060.17411           24.671311           24.671311           24.671311
#> 25 1106.87782           25.575150           25.575150           25.575150
#> 26 1153.58152           26.564310           26.564310           26.564310
#> 27 1200.28523           27.520162           27.520162           27.520162
#> 28 1246.98894           28.454898           28.454898           28.454898
#> 29 1293.69264           29.346979           29.346979           29.346979
#> 30 1340.39635           30.063707           30.063707           30.063707
#>         watershed
#> 1  Antelope Creek
#> 2  Antelope Creek
#> 3  Antelope Creek
#> 4  Antelope Creek
#> 5  Antelope Creek
#> 6  Antelope Creek
#> 7  Antelope Creek
#> 8  Antelope Creek
#> 9  Antelope Creek
#> 10 Antelope Creek
#> 11 Antelope Creek
#> 12 Antelope Creek
#> 13 Antelope Creek
#> 14 Antelope Creek
#> 15 Antelope Creek
#> 16 Antelope Creek
#> 17 Antelope Creek
#> 18 Antelope Creek
#> 19 Antelope Creek
#> 20 Antelope Creek
#> 21 Antelope Creek
#> 22 Antelope Creek
#> 23 Antelope Creek
#> 24 Antelope Creek
#> 25 Antelope Creek
#> 26 Antelope Creek
#> 27 Antelope Creek
#> 28 Antelope Creek
#> 29 Antelope Creek
#> 30 Antelope Creek