Battle Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

battle_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Battle Creek scaling factor of 147% was calculated as the average ratio of Battle Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 5.87 mile low gradient mapped rearing extent in Battle Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 52.4 acres) and floodplain inundation area.

Examples

battle_creek_floodplain

#> # A tibble: 22 x 3
#>    flow_cfs floodplain_acres    watershed
#>       <dbl>            <dbl>        <chr>
#>  1  1133.95             0.00 Battle Creek
#>  2  1281.22             1.81 Battle Creek
#>  3  1428.48             5.31 Battle Creek
#>  4  1575.75             8.25 Battle Creek
#>  5  1723.02            11.41 Battle Creek
#>  6  1870.28            14.20 Battle Creek
#>  7  2017.55            17.18 Battle Creek
#>  8  2164.82            20.00 Battle Creek
#>  9  2312.08            22.72 Battle Creek
#> 10  2459.35            25.05 Battle Creek
#> # ... with 12 more rows