Bear River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

bear_river_floodplain

Format

dataframe with 37 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run Modeling

A 16.40 mile portion of the entire mapped rearing extent of 32.05 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 32.05 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

bear_river_floodplain

#>    flow_cfs FR_floodplain_acres ST_floodplain_acres  watershed
#> 1        10             0.00000             0.00000 Bear River
#> 2       160             0.00000             0.00000 Bear River
#> 3       310             0.00000             0.00000 Bear River
#> 4       460            53.99320            53.99320 Bear River
#> 5       610            93.46025            93.46025 Bear River
#> 6       760           132.92730           132.92730 Bear River
#> 7       910           156.45675           156.45675 Bear River
#> 8       927           159.12342           159.12342 Bear River
#> 9      1060           179.98620           179.98620 Bear River
#> 10     1210           198.92625           198.92625 Bear River
#> 11     1360           214.54888           214.54888 Bear River
#> 12     1510           227.29767           227.29767 Bear River
#> 13     1660           240.04646           240.04646 Bear River
#> 14     1810           252.79525           252.79525 Bear River
#> 15     1960           265.47074           265.47074 Bear River
#> 16     2110           276.24298           276.24298 Bear River
#> 17     2260           287.01523           287.01523 Bear River
#> 18     2410           296.54091           296.54091 Bear River
#> 19     2560           305.32669           305.32669 Bear River
#> 20     2710           317.59032           317.59032 Bear River
#> 21     2860           327.57361           327.57361 Bear River
#> 22     3010           335.81340           335.81340 Bear River
#> 23     3160           345.73694           345.73694 Bear River
#> 24     3310           354.41635           354.41635 Bear River
#> 25     3460           363.09280           363.09280 Bear River
#> 26     3610           372.48949           372.48949 Bear River
#> 27     3760           382.62196           382.62196 Bear River
#> 28     3910           392.03032           392.03032 Bear River
#> 29     4060           406.15697           406.15697 Bear River
#> 30     4210           418.22324           418.22324 Bear River
#> 31     4360           459.49663           459.49663 Bear River
#> 32     4510           478.98101           478.98101 Bear River
#> 33     4660           497.02603           497.02603 Bear River
#> 34     4810           532.39537           532.39537 Bear River
#> 35     4960           550.73505           550.73505 Bear River
#> 36     5110           571.01362           571.01362 Bear River
#> 37     5260           593.02944           593.02944 Bear River