Big Chico Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

big_chico_creek_floodplain

Format

dataframe with 31 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run Modeling

A 8.74 mile portion of the entire mapped rearing extent of 15.05 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 15.05 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Spring Run and Steelhead Modeling

A 8.74 mile portion of the entire mapped rearing extent of 23.84 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 15.05 miles were classified as low gradient and 8.79 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

big_chico_creek_floodplain

#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1        15            0.000000            0.000000            0.000000
#> 2        65            0.000000            0.000000            0.000000
#> 3       115            0.000000            0.000000            0.000000
#> 4       165            0.000000            0.000000            0.000000
#> 5       215            0.000000            0.000000            0.000000
#> 6       265            0.000000            0.000000            0.000000
#> 7       315            0.000000            0.000000            0.000000
#> 8       365            0.000000            0.000000            0.000000
#> 9       415            0.000000            0.000000            0.000000
#> 10      465            0.000000            0.000000            0.000000
#> 11      515            0.000000            0.000000            0.000000
#> 12      565            0.000000            0.000000            0.000000
#> 13      615            0.000000            0.000000            0.000000
#> 14      665            0.000000            0.000000            0.000000
#> 15      715            0.000000            0.000000            0.000000
#> 16      765            0.000000            0.000000            0.000000
#> 17      815            0.000000            0.000000            0.000000
#> 18      865            0.000000            0.000000            0.000000
#> 19      915            0.000000            0.000000            0.000000
#> 20      965            0.000000            0.000000            0.000000
#> 21     1015            0.000000            0.000000            0.000000
#> 22     1065            0.000000            0.000000            0.000000
#> 23     1115            0.000000            0.000000            0.000000
#> 24     1165            0.000000            0.000000            0.000000
#> 25     1215            0.000000            0.000000            0.000000
#> 26     1265            1.605321            1.699004            1.699004
#> 27     1315            4.209542            4.455202            4.455202
#> 28     1365            6.788936            7.185124            7.185124
#> 29     1415           10.011005           10.595226           10.595226
#> 30     1465           13.039603           13.800567           13.800567
#> 31     1515           15.689879           16.605507           16.605507
#>          watershed
#> 1  Big Chico Creek
#> 2  Big Chico Creek
#> 3  Big Chico Creek
#> 4  Big Chico Creek
#> 5  Big Chico Creek
#> 6  Big Chico Creek
#> 7  Big Chico Creek
#> 8  Big Chico Creek
#> 9  Big Chico Creek
#> 10 Big Chico Creek
#> 11 Big Chico Creek
#> 12 Big Chico Creek
#> 13 Big Chico Creek
#> 14 Big Chico Creek
#> 15 Big Chico Creek
#> 16 Big Chico Creek
#> 17 Big Chico Creek
#> 18 Big Chico Creek
#> 19 Big Chico Creek
#> 20 Big Chico Creek
#> 21 Big Chico Creek
#> 22 Big Chico Creek
#> 23 Big Chico Creek
#> 24 Big Chico Creek
#> 25 Big Chico Creek
#> 26 Big Chico Creek
#> 27 Big Chico Creek
#> 28 Big Chico Creek
#> 29 Big Chico Creek
#> 30 Big Chico Creek
#> 31 Big Chico Creek