Butte Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

butte_creek_floodplain

Format

dataframe with 31 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run Modeling

A 17.14 mile portion of the entire mapped rearing extent of 17.14 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 17.14 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Spring Run and Steelhead Modeling

A 17.14 mile portion of the entire mapped rearing extent of 31.67 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 17.14 miles were classified as low gradient and 14.53 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

butte_creek_floodplain

#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1        10             0.00000             0.00000             0.00000
#> 2       110             0.00000             0.00000             0.00000
#> 3       210             0.00000             0.00000             0.00000
#> 4       310             0.00000             0.00000             0.00000
#> 5       410            19.60890            21.27080            21.27080
#> 6       450            31.27005            33.92025            33.92025
#> 7       510            48.76176            52.89442            52.89442
#> 8       610            74.50808            80.82280            80.82280
#> 9       710            91.07548            98.79432            98.79432
#> 10      810           102.17479           110.83431           110.83431
#> 11      910           109.53928           118.82296           118.82296
#> 12     1010           116.26445           126.11810           126.11810
#> 13     1110           126.81067           137.55813           137.55813
#> 14     1210           133.35915           144.66161           144.66161
#> 15     1310           140.57358           152.48748           152.48748
#> 16     1410           145.73412           158.08538           158.08538
#> 17     1510           150.49777           163.25277           163.25277
#> 18     1610           158.01241           171.40429           171.40429
#> 19     1710           163.12795           176.95338           176.95338
#> 20     1810           174.98117           189.81118           189.81118
#> 21     1910           180.39233           195.68096           195.68096
#> 22     2010           184.77699           200.43722           200.43722
#> 23     2110           189.18667           205.22063           205.22063
#> 24     2210           193.98334           210.42383           210.42383
#> 25     2310           200.55972           217.55756           217.55756
#> 26     2410           204.55256           221.88881           221.88881
#> 27     2510           208.56300           226.23914           226.23914
#> 28     2610           214.61182           232.80061           232.80061
#> 29     2710           218.42521           236.93719           236.93719
#> 30     2810           222.64108           241.51037           241.51037
#> 31     2910           226.41997           245.60953           245.60953
#>      watershed
#> 1  Butte Creek
#> 2  Butte Creek
#> 3  Butte Creek
#> 4  Butte Creek
#> 5  Butte Creek
#> 6  Butte Creek
#> 7  Butte Creek
#> 8  Butte Creek
#> 9  Butte Creek
#> 10 Butte Creek
#> 11 Butte Creek
#> 12 Butte Creek
#> 13 Butte Creek
#> 14 Butte Creek
#> 15 Butte Creek
#> 16 Butte Creek
#> 17 Butte Creek
#> 18 Butte Creek
#> 19 Butte Creek
#> 20 Butte Creek
#> 21 Butte Creek
#> 22 Butte Creek
#> 23 Butte Creek
#> 24 Butte Creek
#> 25 Butte Creek
#> 26 Butte Creek
#> 27 Butte Creek
#> 28 Butte Creek
#> 29 Butte Creek
#> 30 Butte Creek
#> 31 Butte Creek