Bypass Flow to Habitat Area Relationships

A dataset containing the suitable habitat area in square meters as a function of flow in cubic feet per second for both instream and floodplain.

yolo_bypass_floodplain

yolo_bypass_instream

sutter_bypass_floodplain

sutter_bypass_instream

Format

4 dataframes, one for each bypass and floodplain or instream habitat type

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

Habitat estimates from NOAA NMFS Winter Run Life Cycle Model

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s

Yolo Bypass

flow_cfs

integer flow value in cubic feet per second

Yolo Bypass 1

suitable floodplain area in square meters in the Yolo Bypass, Fremont Weir to Sacramento Weir

Yolo Bypass 2

suitable floodplain area in square meters in the Yolo Bypass below Sacramento Weir

Sutter Bypass

flow_cfs

integer flow value in cubic feet per second

Sutter Bypass 1

suitable floodplain area in square meters in the Sutter Bypass, to Moulton Weir

Sutter Bypass 2

suitable floodplain area in square meters in the Sutter Bypass, to Colusa Weir

Sutter Bypass 3

suitable floodplain area in square meters in the Sutter Bypass, to Tisdale Weir

Sutter Bypass 4

suitable floodplain area in square meters in the Sutter Bypass below Tisdale Weir

Examples

yolo_bypass_floodplain

#> # A tibble: 2 x 3
#>   flow_cfs `Yolo Bypass 1` `Yolo Bypass 2`
#>      <dbl>           <dbl>           <dbl>
#> 1   175000              0                0
#> 2   200000         132552.               0