Calaveras River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

calaveras_river_floodplain

Format

dataframe with 33 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled TUFLOW hydraulic model with 1D channel and 2D overbank components

Fall Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for the Tuolumne River based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.20 percent of the Tuolumne River values based on the ratio of mean flow from December to June between this watershed and the Tuolumne River. Of the entire mapped 18.48 mile rearing extent in this watershed, 18.48 miles were classifiedas low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

calaveras_river_floodplain

#>      flow_cfs FR_floodplain_acres ST_floodplain_acres       watershed
#> 1    20.04956            0.000000            0.000000 Calaveras River
#> 2    50.12390            0.000000            0.000000 Calaveras River
#> 3   100.24780            0.000000            0.000000 Calaveras River
#> 4   150.37170            0.000000            0.000000 Calaveras River
#> 5   200.49561            0.000000            0.000000 Calaveras River
#> 6   250.61951            2.452816            2.452816 Calaveras River
#> 7   300.74341            5.230114            5.230114 Calaveras River
#> 8   350.86731            7.971828            7.971828 Calaveras River
#> 9   400.99121           10.699097           10.699097 Calaveras River
#> 10  451.11511           13.426719           13.426719 Calaveras River
#> 11  501.23901           16.877290           16.877290 Calaveras River
#> 12  551.36292           20.242601           20.242601 Calaveras River
#> 13  601.48682           23.176327           23.176327 Calaveras River
#> 14  651.61072           26.564715           26.564715 Calaveras River
#> 15  701.73462           29.953103           29.953103 Calaveras River
#> 16  751.85852           33.302913           33.302913 Calaveras River
#> 17  801.98242           36.652722           36.652722 Calaveras River
#> 18  852.10633           39.963248           39.963248 Calaveras River
#> 19  902.23023           43.273775           43.273775 Calaveras River
#> 20  952.35413           46.374322           46.374322 Calaveras River
#> 21 1002.47803           49.474868           49.474868 Calaveras River
#> 22 1052.60193           53.141585           53.141585 Calaveras River
#> 23 1102.72583           56.808302           56.808302 Calaveras River
#> 24 1152.84973           60.286706           60.286706 Calaveras River
#> 25 1202.97364           63.765111           63.765111 Calaveras River
#> 26 1253.09754           67.379156           67.379156 Calaveras River
#> 27 1303.22144           70.993202           70.993202 Calaveras River
#> 28 1353.34534           74.022933           74.022933 Calaveras River
#> 29 1403.46924           77.052665           77.052665 Calaveras River
#> 30 1453.59314           80.274408           80.274408 Calaveras River
#> 31 1503.71704           83.496151           83.496151 Calaveras River
#> 32 1553.84095           86.899865           86.899865 Calaveras River
#> 33 1603.96485           90.303578           90.303578 Calaveras River
#> 34 1654.08875          101.885576          101.885576 Calaveras River
#> 35 1704.21265          113.467574          113.467574 Calaveras River
#> 36 1754.33655          118.350482          118.350482 Calaveras River
#> 37 1804.46045          123.233389          123.233389 Calaveras River