Clear Creek Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

clear_creek_instream

Format

dataframe with 23 rows and 11 variables:

flow_cfs

integer flow value in cubic feet per second

FR_fry_wua

fall run fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

fall run juvenile WUA in square feet per 1000 feet

FR_spawn_wua

fall run spawning WUA in square feet per 1000 feet

SR_fry_wua

spring run fry (up to 50 mm) WUA in square feet per 1000 feet

SR_juv_wua

spring run fry juvenile WUA in square feet per 1000 feet

SR_spawn_wua

spring run fry spawning WUA in square feet per 1000 feet

ST_fry_wua

steelhead fry (up to 50 mm) WUA in square feet per 1000 feet

ST_juv_wua

steelhead juvenile WUA in square feet per 1000 feet

ST_spawn_wua

steelhead spawning WUA in square feet per 1000 feet

watershed

name of watershed

Source

FWS 2007-2013

Details

The spawning and rearing WUA were derived from a FWS River2D model. The WUA values are from three river segments, Upper Alluvial (2.27 miles), Canyon Segment (7.33 miles), and Lower Alluvial (8.81 miles). Fall Run are only in the Lower Alluvial Segment and there is a segregation weir that prevents the Spring Run from spawning in the Lower Alluvial Segment.

Examples

clear_creek_instream

#> # A tibble: 23 x 11
#>    flow_cfs FR_fry_wua FR_juv_wua FR_spawn_wua SR_fry_wua SR_juv_wua
#>       <int>      <dbl>      <dbl>        <dbl>      <dbl>      <dbl>
#>  1       50  11526.287   4835.135     1679.931   3138.600   1862.912
#>  2       75  11367.484   5341.167     2300.416   3094.353   2379.461
#>  3      100  11082.297   5753.491     2798.860   3115.875   2843.080
#>  4      125  10788.468   6089.671     3247.837   3178.732   3259.325
#>  5      150  10549.264   6381.845     3632.021   3192.209   3639.398
#>  6      175  10280.221   6642.073     3995.782   3237.371   4026.180
#>  7      200  10113.615   6840.539     4250.185   3280.250   4334.662
#>  8      225   9967.087   6995.623     4436.612   3398.752   4608.548
#>  9      250   9847.002   7099.027     4566.307   3481.731   4869.358
#> 10      275   9768.621   7196.712     4644.043   3658.718   5092.382
#> # ... with 13 more rows, and 5 more variables: SR_spawn_wua <dbl>,
#> #   ST_fry_wua <dbl>, ST_juv_wua <dbl>, ST_spawn_wua <dbl>, watershed <chr>