Cosumnes River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

cosumnes_river_floodplain

Format

dataframe with 37 rows and 4 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled TUFLOW hydraulic model with 1D channel and 2D overbank components

Fall Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for the Tuolumne River based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.50 percent of the Tuolumne River values based on the ratio of mean flow from December to June between this watershed and the Tuolumne River. Of the entire mapped 41.34 mile rearing extent in this watershed, 41.34 miles were classifiedas low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

cosumnes_river_floodplain

#>     flow_cfs FR_floodplain_acres ST_floodplain_acres      watershed
#> 1    50.2672             0.00000             0.00000 Cosumnes River
#> 2   125.6680             0.00000             0.00000 Cosumnes River
#> 3   251.3360             0.00000             0.00000 Cosumnes River
#> 4   377.0040             0.00000             0.00000 Cosumnes River
#> 5   502.6720             0.00000             0.00000 Cosumnes River
#> 6   628.3400            13.75380            13.75380 Cosumnes River
#> 7   754.0081            29.32708            29.32708 Cosumnes River
#> 8   879.6761            44.70083            44.70083 Cosumnes River
#> 9  1005.3441            59.99358            59.99358 Cosumnes River
#> 10 1131.0121            75.28831            75.28831 Cosumnes River
#> 11 1256.6801            94.63688            94.63688 Cosumnes River
#> 12 1382.3481           113.50736           113.50736 Cosumnes River
#> 13 1508.0161           129.95778           129.95778 Cosumnes River
#> 14 1633.6841           148.95766           148.95766 Cosumnes River
#> 15 1759.3521           167.95754           167.95754 Cosumnes River
#> 16 1885.0201           186.74109           186.74109 Cosumnes River
#> 17 2010.6881           205.52465           205.52465 Cosumnes River
#> 18 2136.3562           224.08793           224.08793 Cosumnes River
#> 19 2262.0242           242.65121           242.65121 Cosumnes River
#> 20 2387.6922           260.03706           260.03706 Cosumnes River
#> 21 2513.3602           277.42291           277.42291 Cosumnes River
#> 22 2639.0282           297.98348           297.98348 Cosumnes River
#> 23 2764.6962           318.54404           318.54404 Cosumnes River
#> 24 2890.3642           338.04867           338.04867 Cosumnes River
#> 25 3016.0322           357.55331           357.55331 Cosumnes River
#> 26 3141.7002           377.81853           377.81853 Cosumnes River
#> 27 3267.3682           398.08375           398.08375 Cosumnes River
#> 28 3393.0362           415.07251           415.07251 Cosumnes River
#> 29 3518.7043           432.06128           432.06128 Cosumnes River
#> 30 3644.3723           450.12672           450.12672 Cosumnes River
#> 31 3770.0403           468.19216           468.19216 Cosumnes River
#> 32 3895.7083           487.27797           487.27797 Cosumnes River
#> 33 4021.3763           506.36379           506.36379 Cosumnes River
#> 34 4147.0443           571.30811           571.30811 Cosumnes River
#> 35 4272.7123           636.25243           636.25243 Cosumnes River
#> 36 4398.3803           663.63260           663.63260 Cosumnes River
#> 37 4524.0483           691.01277           691.01277 Cosumnes River