Cosumnes River Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

cosumnes_river_instream

Format

dataframe with 14 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Sadie Gill

Details

The Cosumnes River instream rearing habitat has not been modeled. The WUA values were estimated using the mean WUA at each flow from Mokelumne and Calaveras Rivers.

Examples

cosumnes_river_instream

#> # A tibble: 10 x 5
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua      watershed
#>       <dbl>        <dbl>      <dbl>      <dbl>          <chr>
#>  1      100     1424.717  12662.125  12973.250 Cosumnes River
#>  2      200     1406.727   9528.625  10806.500 Cosumnes River
#>  3      300     1380.998   7357.625  10206.250 Cosumnes River
#>  4      400     1705.270   6000.500   8725.000 Cosumnes River
#>  5      500     1804.686   4543.250   6861.750 Cosumnes River
#>  6      600     1804.686   4051.625   5911.000 Cosumnes River
#>  7      700     1804.686   3821.500   4921.625 Cosumnes River
#>  8      800     1804.686   3649.250   4800.250 Cosumnes River
#>  9      900     1804.686   3525.875   4424.125 Cosumnes River
#> 10     1000     1804.686   3220.625   4082.000 Cosumnes River