Cottonwood Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

cottonwood_creek_floodplain

Format

dataframe with 10 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Steelhead when needed)

watershed

name of watershed

Source

FWS

Details

The habitat area was derived from the FWS 2017 HEC-RAS 1D model.

Examples

cottonwood_creek_floodplain

#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1        40            0.000000            0.000000            0.000000
#> 2       400            0.000000            0.000000            0.000000
#> 3       760            0.000000            0.000000            0.000000
#> 4      1120            0.000000            0.000000            0.000000
#> 5      1480            0.000000            0.000000            0.000000
#> 6      1840            0.000000            0.000000            0.000000
#> 7      1870            0.000000            0.000000            0.000000
#> 8      2200            0.000000            0.000000            0.000000
#> 9      2560            0.000000            0.000000            0.000000
#> 10     2920            0.000000            0.000000            0.000000
#> 11     3280            0.000000            0.000000            0.000000
#> 12     3640            0.000000            0.000000            0.000000
#> 13     4000            0.000000            0.000000            0.000000
#> 14     4360            0.000000            0.000000            0.000000
#> 15     4720            0.000000            0.000000            0.000000
#> 16     5080            0.000000            0.000000            0.000000
#> 17     5440            0.000000            0.000000            0.000000
#> 18     5800            0.000000            0.000000            0.000000
#> 19     6160            0.000000            0.000000            0.000000
#> 20     6520            0.000000            0.000000            0.000000
#> 21     6880            0.000000            0.000000            0.000000
#> 22     7240            0.000000            0.000000            0.000000
#> 23     7600            0.000000            0.000000            0.000000
#> 24     7960            5.910398            8.477617            9.079739
#> 25     8320           13.261246           19.021353           20.372342
#> 26     8680           19.264434           27.632064           29.594628
#> 27     9040           24.203451           34.716375           37.182101
#> 28     9400           29.180553           41.855313           44.828082
#> 29     9760           33.393572           47.898284           51.300254
#> 30    10120           38.032135           54.551636           58.426160
#> 31    10480           43.174776           61.928017           66.326447
#> 32    10840           47.516971           68.156272           72.997063
#>           watershed
#> 1  Cottonwood Creek
#> 2  Cottonwood Creek
#> 3  Cottonwood Creek
#> 4  Cottonwood Creek
#> 5  Cottonwood Creek
#> 6  Cottonwood Creek
#> 7  Cottonwood Creek
#> 8  Cottonwood Creek
#> 9  Cottonwood Creek
#> 10 Cottonwood Creek
#> 11 Cottonwood Creek
#> 12 Cottonwood Creek
#> 13 Cottonwood Creek
#> 14 Cottonwood Creek
#> 15 Cottonwood Creek
#> 16 Cottonwood Creek
#> 17 Cottonwood Creek
#> 18 Cottonwood Creek
#> 19 Cottonwood Creek
#> 20 Cottonwood Creek
#> 21 Cottonwood Creek
#> 22 Cottonwood Creek
#> 23 Cottonwood Creek
#> 24 Cottonwood Creek
#> 25 Cottonwood Creek
#> 26 Cottonwood Creek
#> 27 Cottonwood Creek
#> 28 Cottonwood Creek
#> 29 Cottonwood Creek
#> 30 Cottonwood Creek
#> 31 Cottonwood Creek
#> 32 Cottonwood Creek