Cow Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

cow_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Cow Creek scaling factor of 216% was calculated as the average ratio of Cow Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 8.6 mile low gradient and 55.6 miles of high gradient mapped rearing extent in Cow Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 217.5 acres) and floodplain inundation area.

Examples

cow_creek_floodplain

#> # A tibble: 22 x 3
#>    flow_cfs floodplain_acres watershed
#>       <dbl>            <dbl> <chr>    
#>  1     1660             0    Cow Creek
#>  2     1876             6.38 Cow Creek
#>  3     2092            18.8  Cow Creek
#>  4     2307            29.1  Cow Creek
#>  5     2523            40.3  Cow Creek
#>  6     2738            50.1  Cow Creek
#>  7     2954            60.6  Cow Creek
#>  8     3170            70.6  Cow Creek
#>  9     3385            80.2  Cow Creek
#> 10     3601            88.4  Cow Creek
#> # ... with 12 more rows