Cow Creek Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

cow_creek_instream

Format

dataframe with 30 rows and 4 variables:

flow_cfs

integer flow value in cubic feet per second

FR_fry_wua

fall run fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

fall run juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

FWS 2011

Details

The rearing WUA values are from a River2D model done on South Cow Creek. The values are from two river segments, Valley Floor Reach (5.11 miles) and Boero Reach (1.68 miles).

Examples

cow_creek_instream

#> # A tibble: 30 x 4
#>    flow_cfs FR_fry_wua FR_juv_wua watershed
#>       <int>      <dbl>      <dbl> <chr>    
#>  1       10     12208.      3748. Cow Creek
#>  2       20     11880.      4674. Cow Creek
#>  3       30     12324.      5032. Cow Creek
#>  4       40     12859.      5662. Cow Creek
#>  5       50     13410.      6190. Cow Creek
#>  6       60     14124.      6631. Cow Creek
#>  7       70     14426.      7045. Cow Creek
#>  8       80     14535.      7420. Cow Creek
#>  9       90     14664.      7665. Cow Creek
#> 10      100     14662.      7952. Cow Creek
#> # ... with 20 more rows