Deer Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

deer_creek_floodplain

Format

dataframe with 30 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Steelhead when needed)

watershed

name of watershed

Source

Deer Creek Watershed Conservancy (DCWC) 1D HEC-RAS model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. Model extent was from the confluence with the Sacramento River upstream 11.8 miles to the Deer Creek Irrigation District diversion dam. Floodplain area for the remaining 6.6 miles in the mapped rearing extent was calculated by determing the floodplain area per river mile in the modeled reach and applying this ratio to the unmodeled reach with a reduction factor of 10% to account for the transition from low gradient to high gradient geomorphic conditions in the upper 6.6 miles of rearing extent. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 58.7 acres) and floodplain inundation area.

Examples

deer_creek_floodplain

#> # A tibble: 22 x 3
#>    flow_cfs floodplain_acres watershed 
#>       <int>            <dbl> <chr>     
#>  1      770             0    Deer Creek
#>  2      870             2.60 Deer Creek
#>  3      970             7.63 Deer Creek
#>  4     1070            11.9  Deer Creek
#>  5     1170            16.4  Deer Creek
#>  6     1270            20.4  Deer Creek
#>  7     1370            24.7  Deer Creek
#>  8     1470            28.7  Deer Creek
#>  9     1570            32.6  Deer Creek
#> 10     1670            36.0  Deer Creek
#> # ... with 12 more rows