Delta Habitat Area

A dataset containing the area of highly suitable habitat within the North and South Deltas

delta_habitat

Format

dataframe with 372 rows and 3 variables

date

modeled results for 1980-2010

North Delta

high quality habitat area in square meters

South Delta

high quality habitat area in square meters

Source

Modeling Output: Correigh Green correigh.greene@noaa.gov

Delta Node Selection: Mark Tompkins mtompkins@flowwest.com

Data Wrangling: Sadie Gill sgill@flowwest.com

Details

Habitat estimates from Correigh Greene's Winter Run Life Cycle Model. The model outputs habitat area estimates for months December-May. Habitat for months June-Novemeber are represented by the yearly average habitat.

High quality habitat defined by:

Channel type: mainstem, distributaries, or open water

Depth: > 0.2 m and <= 1.5 m

Cover Vegetated edge