Feather River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

feather_river_floodplain

Format

dataframe with 40 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run, Spring Run and Steelhead Modeling

The entire mapped rearing extent of 64.98 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Active channel area of 2363.20 acres estimated through remote sensing analysis was subtracted from total inundated area to get inundated floodplain area.

Examples

feather_river_floodplain

#> # A tibble: 40 x 5
#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#>       <dbl>               <dbl>               <dbl>               <dbl>
#>  1     1100              0.0000              0.0000              0.0000
#>  2     2600            471.5575            471.5575            471.5575
#>  3     4056            974.3999            974.3999            974.3999
#>  4     4100            989.5956            989.5956            989.5956
#>  5     5600           1174.1560           1174.1560           1174.1560
#>  6     6983           1299.3545           1299.3545           1299.3545
#>  7     7100           1309.9461           1309.9461           1309.9461
#>  8     8600           1477.9909           1477.9909           1477.9909
#>  9    10100           1648.3905           1648.3905           1648.3905
#> 10    11600           1773.4152           1773.4152           1773.4152
#> # ... with 30 more rows, and 1 more variables: watershed <chr>