Habitat Data

Datasets used to calculate spawning, instream, and floodplain rearing habitat areas for Chinook Salmon and Steelhead

Instream

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - flow value in cubic feet per second

FR_spawn_wua - Spawning WUA in square feet per 1000 feet

FR_fry_wua - Fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua - Juvenile WUA in square feet per 1000 feet

watershed - Name of watershed

Floodplain

Datasets containing the total inuncated area in acres as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - Flow value in cubic feet per second

FR_floodplain_acres - Total inundated area in acres for Fall Run

SR_floodplain_acres - Total inundated area in acres for Spring Run

ST_floodplain_acres - Total inundated area in acres for Steelhead

watershed - Name of watershed

Modeling Details for Streams

American River

Antelope Creek

Battle Creek

Bear Creek

Bear River

Big Chico Creek

Butte Creek

Calaveras River

Clear Creek

Cosumnes River

Cottonwood Creek

Cow Creek

Deer Creek

Elder Creek

Feather River

Merced River

Mill Creek

Mokelumne River

Paynes Creek

Sacramento River

San Joaquin River

Stanislaus River

Stony Creek

Thomes Creek

Tuolumne River

Yuba River

Bypasses

Sutter and Yolo Bypass

Delta

North and South Delta