Lower-mid Sacramento River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

lower_mid_sacramento_river_floodplain

Format

dataframe with 240 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

watershed

name of watershed

Source

missing

Details

The habitat area was derived from a HEC-RAS 1D model. The CVPIA Lower-mid Sacramento River extends from Wilkins Slough to the American River (58.0 mi). This reach overlaps with two of the study's, Battle Creek to the confluence with the Feather River (189.1 mi) and the confluence with the Feather River to Freeport (33.4 mi). To scale the study's results to CVPIA's extents, we calculate the proportion of the Lower-mid Sacramento River above (34%) and below (66%) the Feather River. The floodplain acerage at a given flow is the weighted average of the floodplain area per river mile within each study.

Examples

lower_mid_sacramento_river_floodplain

#> # A tibble: 240 x 3
#>    flow_cfs FR_floodplain_acres                  watershed
#>       <dbl>               <dbl>                      <chr>
#>  1 2637.426             0.00000 Lower-mid Sacramento River
#>  2 2967.296            40.87832 Lower-mid Sacramento River
#>  3 2992.467            40.95891 Lower-mid Sacramento River
#>  4 3384.049            42.21263 Lower-mid Sacramento River
#>  5 3479.393            42.95670 Lower-mid Sacramento River
#>  6 3550.317            43.51020 Lower-mid Sacramento River
#>  7 3854.895            45.56907 Lower-mid Sacramento River
#>  8 3897.321            45.56907 Lower-mid Sacramento River
#>  9 4144.839            45.56907 Lower-mid Sacramento River
#> 10 4355.361            46.54123 Lower-mid Sacramento River
#> # ... with 230 more rows