Merced River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

merced_river_floodplain

Format

dataframe with 23 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage (use for Steelhead when required)

watershed

name of watershed

Source

Modeling: CVFPP2012

Details

The habitat area was derived from the CVFPP 2012 HEC-RAS 1D model. The study only represents the lower 25.5 miles of the Merced River. The modeled results were scaled up to represent the whole rearing region. We calculate area per river mile from modeled reach and assign the product of the river mile distance and floodplain area per river mile for each reach defined as either in the Valley Lowland area, in a higher gradient / laterally confined area between Valley Lowland areas, or in the dredge tailing area. The study area is defined in the CV Habitat Exchange as being within the Valley Lowland, all other reaches with that same category are given the full area rate. River miles outside the Valley Lowland category recieve some proportion of the area rate, detailed below:

RM 0 - 25.5 use areas defined by study

RM 25.5 - 31.2 "Valley Lowland" use 1X area/RM from modeled reach.

RM 31.2 - 43.6 not "Valley Lowland", use 0.5X area/RM relationship from modeled reach.

RM 43.6 - 47.0 "Valley Lowland", use 1X area/RM from modeled reach.

RM 47.0 - 52.0 not "Valley Lowland" and in dredge tailings, use 0.1X area/RM from modeled reach.

Examples

merced_river_floodplain

Scaling Criteria developed by Mark Tompkins \email{mtompkins@flowwest.com}

#> Error: <text>:3:9: unexpected symbol
#> 2: 
#> 3: Scaling Criteria
#>            ^