Mokelumne River to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

mokelumne_river_floodplain

Format

dataframe with 33 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from Tuolumne River flow to floodplain area relationship generated with a TUFLOW 2D hydraulic model

Details

TUFLOW model runs to generate Tuolumne River CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Mokelumne River scaling factor of 50% was calculated as the average ratio of Mokelumne River monthly mean flow to Tuolumne River monthly mean flow between December and June. This scaling factor was applied to Tuolumne River flows and corresponding floodplain area per river mile values, which were then multiplied by the 58.4 mile mapped rearing extent in the Mokelumne River. Remote sensing analysis of aerial photography was used to confirm appropriate scale of instream (active channel area of 241.2 acres) versus floodplain inundation areas.

Examples

mokelumne_river_floodplain

#> # A tibble: 33 x 3
#>    flow_cfs floodplain_acres       watershed
#>       <dbl>            <dbl>           <chr>
#>  1    502.7              0.0 Mokelumne River
#>  2    628.3             19.4 Mokelumne River
#>  3    754.0             41.5 Mokelumne River
#>  4    879.7             63.2 Mokelumne River
#>  5   1005.3             84.8 Mokelumne River
#>  6   1131.0            106.4 Mokelumne River
#>  7   1256.7            133.8 Mokelumne River
#>  8   1382.3            160.4 Mokelumne River
#>  9   1508.0            183.7 Mokelumne River
#> 10   1633.7            210.6 Mokelumne River
#> # ... with 23 more rows