North Delta Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

north_delta_floodplain

Format

dataframe with 73 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

NMFS life cycle modelHendirx 2017 HEC-RAS 1D model

Details

The habitat area was derived from the NMFS life cycle model, Hendirx 2017 HEC-RAS 1D model.

Examples

north_delta_floodplain

#> # A tibble: 73 x 3
#>    flow_cfs FR_floodplain_acres   watershed
#>       <dbl>               <dbl>       <chr>
#>  1 3755.530            0.000000 North Delta
#>  2 7237.521            0.000000 North Delta
#>  3 7420.305            2.990095 North Delta
#>  4 7605.674            2.990095 North Delta
#>  5 8137.583            2.990095 North Delta
#>  6 8629.939            2.990095 North Delta
#>  7 8846.635            5.393880 North Delta
#>  8 9091.533            5.393880 North Delta
#>  9 9403.642            5.940566 North Delta
#> 10 9650.198            9.947793 North Delta
#> # ... with 63 more rows