Pool Area

A dataset containing estimated area of pools within a watershed

pools

Format

dataframe with 31 rows and 3 variables

watershed

CVPIA watershed

SR_pools_sq_meters

Spring Run pool area estimate in square meters

ST_pools_sq_meters

Steelhead pool area estimate in square meters

Source

Percent Pool Estimates: Mark Gard Mark.Gard@Wildlife.ca.gov

Channel Area Estimates: Sadie Gill sgill@flowwest.com

QA/QC: Mark Tompkins mtompkins@flowwest.com

Details

Estimates of percent pools were provided by Mark Gard for some watersheds. The mean percent pools, excluding Feather River, is 23.1%. This value is used for watersheds without an estimate. The total area of pools is then caluculated by multiplying the estimated percent pools by the total channel area. Channel areas were estimated using Google Earth Engine.

Watersheds With Estimated Percent Pools:

Antelope Creek

Battle Creek

Butte Creek

Clear Creek

Cottonwood Creek

Deer Creek

Feather River (*80

Mill Creek

Mokelumne River

Stanislaus River

Tuolumne River

Upper Sacramento River

Yuba River

Note: The area of pools that covers both the Upper Sacramento and Upper-mid Sacramento DSM segments is assigned to the Upper Sacramento River.