Mainstem Sacramento River Instream Rearing Habitat

A dataset containing the instream rearing habitat area in square meters as a function of flow in cubic feet per second

upper_sacramento_river_instream

upper_mid_sacramento_river_instream

lower_mid_sacramento_river_instream

lower_sacramento_river_instream

Format

dataframe with 81 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

rearing_sq_meters

suitable instream rearing area in square meters

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Fall Run and Steelhead Modeling

The entire mapped rearing extent was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model. The high quality depth and high quality velocity ("Pref11") "BankArea" result was used as floodplain area. High quality velocities were assumed to be less than or equal to 0.15 meters per second, and high quality depths were assumed to be between 0.2 meters and 1.5 meters.

Model Scaling

The study's results were scaled using the proportion of the CVPIA reach with each study reach extent.

Upper Sacramento River

The CVPIA Upper Sacramento River extends from Keswick to Red Bluff (59.3 mi). This reach overlaps with two of the study's, Keswick to Battle Creek (28.9 mi) and Battle Creek to the confluence with the Feather River (186.5 mi).

Upper-mid Sacramento River

The study's extent is from Battle Creek to the confluence with the Feather River (186.5 mi). The CVPIA Upper-mid Sacramento River extends from Red Blurr to Wilkins Slough (122.3 mi).

Lower-mid Sacramento River

The CVPIA Lower-mid Sacramento River extends from Wilkins Slough to the American River (58.0 mi). This reach overlaps with two of the study's, Battle Creek to the confluence with the Feather River (186.5 mi) and the confluence with the Feather River to Freeport (33.9 mi).

Lower Sacramento River

The study's extent is from the confluence with the Feather River to Freeport (33.4 mi). The CVPIA Lower Sacramento River extends from the confluence with the American River to Freeport (13.7 mi).

Examples

upper_sacramento_river_instream

#> # A tibble: 81 x 3
#>    flow_cfs rearing_sq_meters              watershed
#>       <int>             <dbl>                  <chr>
#>  1     2000          408108.4 Upper Sacramento River
#>  2     2500          328336.6 Upper Sacramento River
#>  3     3000          300066.5 Upper Sacramento River
#>  4     3500          256124.3 Upper Sacramento River
#>  5     4000          296282.6 Upper Sacramento River
#>  6     4500          295832.9 Upper Sacramento River
#>  7     5000          262956.6 Upper Sacramento River
#>  8     5500          269656.5 Upper Sacramento River
#>  9     6000          260590.7 Upper Sacramento River
#> 10     6500          246153.8 Upper Sacramento River
#> # ... with 71 more rows