Mainstem Sacramento River Habitat Data

Datasets used to calculate spawning, instream, and floodplain rearing habitat areas for Chinook Salmon and Steelhead

Source

Sacramento River Modeling Details

Spawning

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second for Fall Run (FR), Late-fall Run (LFR), Winter Run (WR), and Steelhead (ST).

upper_sac_ACID_boards_in

upper_sac_ACID_boards_out

Data Format:

flow_cfs - flow value in cubic feet per second

FR_spawn_wua - Spawning WUA in square feet per 1000 feet (Fall Run)

LFR_spawn_wua - Spawning WUA in square feet per 1000 feet (Late-Fall Run)

ST_spawn_wua - Spawning WUA in square feet per 1000 feet (Steelhead)

WR_spawn_wua - Spawning WUA in square feet per 1000 feet (Winter Run)

watershed - Name of watershed

Instream

Datasets containing the instream rearing habitat area in square meters as a function of flow in cubic feet per second.

upper_sacramento_river_instream

upper_mid_sacramento_river_instream

lower_mid_sacramento_river_instream

lower_sacramento_river_instream

Data Format:

flow_cfs - Flow value in cubic feet per second

rearing_sq_meters - Suitable rearing area in square meteres

watershed - Name of watershed

Floodplain

Datasets containing the floodplain habitat area in square meters as a function of flow in cubic feet per second.

upper_sacramento_river_floodplain

upper_mid_sacramento_river_floodplain

lower_mid_sacramento_river_floodplain

lower_sacramento_river_floodplain

Data Format:

flow_cfs - Flow value in cubic feet per second

floodplain_sq_meters - Suitable rearing area in square meteres

watershed - Name of watershed