San Joaquin River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

san_joaquin_river_floodplain

Format

dataframe with 32 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run, Spring Run and Steelhead Modeling

The entire mapped rearing extent of 45.68 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Active channel area of 534.20 acres estimated through remote sensing analysis was subtracted from total inundated area to get inundated floodplain area.

Examples

san_joaquin_river_floodplain

#> # A tibble: 32 x 5
#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#>       <dbl>               <dbl>               <dbl>               <dbl>
#>  1       20              0.0000              0.0000              0.0000
#>  2      770              0.0000              0.0000              0.0000
#>  3     1520             94.4063             94.4063             94.4063
#>  4     2214            279.4574            279.4574            279.4574
#>  5     2270            294.3895            294.3895            294.3895
#>  6     3020            702.2399            702.2399            702.2399
#>  7     3770            839.2114            839.2114            839.2114
#>  8     4520           1012.3999           1012.3999           1012.3999
#>  9     5270           1144.9671           1144.9671           1144.9671
#> 10     6020           1301.8202           1301.8202           1301.8202
#> # ... with 22 more rows, and 1 more variables: watershed <chr>