San Joaqin River Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

san_joaquin_river_instream

Format

dataframe with 14 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Sadie Gill

Details

The Lower San Joaquin River instream rearing habitat has not been modeled. The WUA values were estimated using the mean WUA at each flow from the Merced, Stanislaus, and Tuolumne Rivers.

Examples

san_joaquin_river_instream

#> # A tibble: 47 x 5
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua         watershed
#>       <int>        <dbl>      <dbl>      <dbl>             <chr>
#>  1       25     3064.559   28359.27   28359.27 San Joaquin River
#>  2       50     3289.612   28359.27   28359.27 San Joaquin River
#>  3      100     6481.759   29605.21   29605.21 San Joaquin River
#>  4      150     9668.076   29752.65   29752.65 San Joaquin River
#>  5      200    11340.041   28765.61   28765.61 San Joaquin River
#>  6      250    11977.774   27499.19   27499.19 San Joaquin River
#>  7      300    12192.079   26287.65   26287.65 San Joaquin River
#>  8      350    11534.828   25069.58   25069.58 San Joaquin River
#>  9      400    10652.446   23933.09   23933.09 San Joaquin River
#> 10      450     9739.564   22928.94   22928.94 San Joaquin River
#> # ... with 37 more rows