Set Floodplain Habitat Area

This function returns an estimated floodplain area based on watershed, species and flow.

set_floodplain_habitat(watershed, species, flow, flow2 = NULL)

Arguments

watershed
a watershed defined for the SIT model

species
one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)

flow
a flow value in cubic feet per second

Value

floodplain habitat value in square meters

Lower-mid Sacramento River

The Lower-mid Sacramento River has two nodes, one above Fremont Weir (C134) and one below (C160). When calculating habitat for the Lower-Mid Sacramento river, calculate the habitat at each flow node and sum them proportion to the length of stream above and below the weir:

35.6/58 * (habitat at C134) + 22.4/58 * (habitat at C160)

Regional Approximation: When a watershed has no associated floodplain modeling, an approximation is made. [insert method description]

Examples

# floodplain habitat value in square meters for Fall Run in the American River
set_floodplain_habitat("American River", "fr", 34652)

#> [1] 4177973

# floodplain habitat value in square meters for Steelhead in the American River
set_floodplain_habitat("American River", "st", 34652)

#> [1] 4177973