Stanislaus River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

stanislaus_river_floodplain

Format

dataframe with 15 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

SRH-2D hyraulic model developed by NewFields (now FlowWest www.flowwest.com)

Details

Model extent was from the confluence of the San Joaquin River upstream to Goodwin Dam. Active channel area delineated based on geomorphic features to differentiate between instream and floodplain habitat.

Examples

stanislaus_river_floodplain

#> # A tibble: 15 x 3
#>    flow_cfs floodplain_acres watershed       
#>       <int>            <dbl> <chr>           
#>  1      250             0    Stanislaus River
#>  2      500             3.31 Stanislaus River
#>  3      750             7.55 Stanislaus River
#>  4     1000            12.4  Stanislaus River
#>  5     1250            40.5  Stanislaus River
#>  6     1500           104    Stanislaus River
#>  7     1750           150    Stanislaus River
#>  8     2000           183    Stanislaus River
#>  9     2001           183    Stanislaus River
#> 10     2500           257    Stanislaus River
#> 11     3000           365    Stanislaus River
#> 12     3500           420    Stanislaus River
#> 13     4000           503    Stanislaus River
#> 14     4500           623    Stanislaus River
#> 15     5000           810    Stanislaus River