Stony Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

stony_creek_floodplain

Format

dataframe with 32 rows and 4 variables:

flow_cfs

flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled from a Cottonwood Creek flow to floodplain area relationship generated with a hybrid USFWS / FEMA 1D HEC-RAS hydraulic model

Fall Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.59 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 26.12 mile rearing extent in this watershed, 26.12 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

stony_creek_floodplain

#>     flow_cfs FR_floodplain_acres ST_floodplain_acres   watershed
#> 1    23.4904            0.000000            0.000000 Stony Creek
#> 2   234.9040            0.000000            0.000000 Stony Creek
#> 3   446.3176            0.000000            0.000000 Stony Creek
#> 4   657.7311            0.000000            0.000000 Stony Creek
#> 5   869.1447            0.000000            0.000000 Stony Creek
#> 6  1080.5583            0.000000            0.000000 Stony Creek
#> 7  1098.1761            0.000000            0.000000 Stony Creek
#> 8  1291.9719            0.000000            0.000000 Stony Creek
#> 9  1503.3854            0.000000            0.000000 Stony Creek
#> 10 1714.7990            0.000000            0.000000 Stony Creek
#> 11 1926.2126            0.000000            0.000000 Stony Creek
#> 12 2137.6262            0.000000            0.000000 Stony Creek
#> 13 2349.0398            0.000000            0.000000 Stony Creek
#> 14 2560.4533            0.000000            0.000000 Stony Creek
#> 15 2771.8669            0.000000            0.000000 Stony Creek
#> 16 2983.2805            0.000000            0.000000 Stony Creek
#> 17 3194.6941            0.000000            0.000000 Stony Creek
#> 18 3406.1077            0.000000            0.000000 Stony Creek
#> 19 3617.5212            0.000000            0.000000 Stony Creek
#> 20 3828.9348            0.000000            0.000000 Stony Creek
#> 21 4040.3484            0.000000            0.000000 Stony Creek
#> 22 4251.7620            0.000000            0.000000 Stony Creek
#> 23 4463.1755            0.000000            0.000000 Stony Creek
#> 24 4674.5891            4.054819            4.054819 Stony Creek
#> 25 4886.0027            9.097857            9.097857 Stony Creek
#> 26 5097.4163           13.216335           13.216335 Stony Creek
#> 27 5308.8299           16.604740           16.604740 Stony Creek
#> 28 5520.2434           20.019274           20.019274 Stony Creek
#> 29 5731.6570           22.909609           22.909609 Stony Creek
#> 30 5943.0706           26.091888           26.091888 Stony Creek
#> 31 6154.4842           29.619989           29.619989 Stony Creek
#> 32 6365.8978           32.598945           32.598945 Stony Creek