Sutter Bypass Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

sutter_bypass_floodplain

Format

dataframe with 41 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

Sutter Bypass 1

floodplain acreage in the Sutter Bypass, to Moulton Weir

Sutter Bypass 2

floodplain acreage in the Sutter Bypass, to Colusa Weir

Sutter Bypass 3

floodplain acreage in the Sutter Bypass, to Tisdale Weir

Sutter Bypass 4

floodplain acreage in the Sutter Bypass below Tisdale Weir

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

Habitat estimates are freom the Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model.

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s

Examples

sutter_bypass_floodplain

#> # A tibble: 41 x 5
#>    flow_cfs `Sutter Bypass 1` `Sutter Bypass 2` `Sutter Bypass 3`
#>  *    <dbl>             <dbl>             <dbl>             <dbl>
#>  1       10         0.0000000          6.737225         0.0000000
#>  2       20         0.0000000         26.480142         0.6348337
#>  3       30         0.1919283         78.072729         0.6560252
#>  4       40         0.2002831         99.774472         0.6740380
#>  5       50         0.2077097        110.380948         0.6899316
#>  6       60         1.1788731        109.903506         0.7037061
#>  7       70         1.1856209        188.665664         0.7164209
#>  8       80         1.8673173        194.535464         0.7291358
#>  9       90         1.8684968        201.806997         0.7407911
#> 10      100         1.8696763        248.119919         0.7503273
#> # ... with 31 more rows, and 1 more variables: `Sutter Bypass 4` <dbl>