Sutter Bypass Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet as a function of flow in cubic feet per second

sutter_bypass_instream

Format

dataframe with 41 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

Sutter Bypass 1

juvenile rearing habitat WUA in the Sutter Bypass, to Moulton Weir

Sutter Bypass 2

juvenile rearing habitat WUA in the Sutter Bypass, to Colusa Weir

Sutter Bypass 3

juvenile rearing habitat WUA in the Sutter Bypass, to Tisdale Weir

Sutter Bypass 4

juvenile rearing habitat WUA in the Sutter Bypass below Tisdale Weir

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

Habitat estimates from Correigh Greene's Winter Run Life Cycle Model

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s

Examples

sutter_bypass_instream

#> # A tibble: 41 x 5
#>    flow_cfs `Sutter Bypass 1` `Sutter Bypass 2` `Sutter Bypass 3`
#>  *    <dbl>             <dbl>             <dbl>             <dbl>
#>  1       10           2613292           1168074           5674414
#>  2       20           5302250           2632234           7529966
#>  3       30           6221161           3808180           9582359
#>  4       40           6703418           4275595          12193654
#>  5       50           7498144           4422091          14231994
#>  6       60          11703359           4667249          15132002
#>  7       70          12254095           4825101          15538677
#>  8       80          12193444           5152135          16445649
#>  9       90          12118327           5062672          17111271
#> 10      100          11689165           5122701          18389934
#> # ... with 31 more rows, and 1 more variables: `Sutter Bypass 4` <dbl>