Thomes Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

thomes_creek_floodplain

Format

dataframe with 32 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled from a Cottonwood Creek flow to floodplain area relationship generated with a hybrid USFWS / FEMA 1D HEC-RAS hydraulic model

Fall Run Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.37 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 37.47 mile rearing extent in this watershed, 37.47 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Spring Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.37 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 42.70 mile rearing extent in this watershed, 37.47 miles were classified as low gradient and 5.22 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

thomes_creek_floodplain

#>      flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1    14.71286            0.000000            0.000000            0.000000
#> 2   147.12859            0.000000            0.000000            0.000000
#> 3   279.54432            0.000000            0.000000            0.000000
#> 4   411.96005            0.000000            0.000000            0.000000
#> 5   544.37578            0.000000            0.000000            0.000000
#> 6   676.79151            0.000000            0.000000            0.000000
#> 7   687.82615            0.000000            0.000000            0.000000
#> 8   809.20724            0.000000            0.000000            0.000000
#> 9   941.62297            0.000000            0.000000            0.000000
#> 10 1074.03870            0.000000            0.000000            0.000000
#> 11 1206.45443            0.000000            0.000000            0.000000
#> 12 1338.87016            0.000000            0.000000            0.000000
#> 13 1471.28589            0.000000            0.000000            0.000000
#> 14 1603.70162            0.000000            0.000000            0.000000
#> 15 1736.11735            0.000000            0.000000            0.000000
#> 16 1868.53308            0.000000            0.000000            0.000000
#> 17 2000.94880            0.000000            0.000000            0.000000
#> 18 2133.36453            0.000000            0.000000            0.000000
#> 19 2265.78026            0.000000            0.000000            0.000000
#> 20 2398.19599            0.000000            0.000000            0.000000
#> 21 2530.61172            0.000000            0.000000            0.000000
#> 22 2663.02745            0.000000            0.000000            0.000000
#> 23 2795.44318            0.000000            0.000000            0.000000
#> 24 2927.85891            3.643509            3.694287            3.694287
#> 25 3060.27464            8.174995            8.288925            8.288925
#> 26 3192.69037           11.875705           12.041210           12.041210
#> 27 3325.10610           14.920400           15.128336           15.128336
#> 28 3457.52183           17.988572           18.239268           18.239268
#> 29 3589.93756           20.585719           20.872610           20.872610
#> 30 3722.35329           23.445197           23.771938           23.771938
#> 31 3854.76902           26.615417           26.986339           26.986339
#> 32 3987.18475           29.292195           29.700423           29.700423
#>       watershed
#> 1  Thomes Creek
#> 2  Thomes Creek
#> 3  Thomes Creek
#> 4  Thomes Creek
#> 5  Thomes Creek
#> 6  Thomes Creek
#> 7  Thomes Creek
#> 8  Thomes Creek
#> 9  Thomes Creek
#> 10 Thomes Creek
#> 11 Thomes Creek
#> 12 Thomes Creek
#> 13 Thomes Creek
#> 14 Thomes Creek
#> 15 Thomes Creek
#> 16 Thomes Creek
#> 17 Thomes Creek
#> 18 Thomes Creek
#> 19 Thomes Creek
#> 20 Thomes Creek
#> 21 Thomes Creek
#> 22 Thomes Creek
#> 23 Thomes Creek
#> 24 Thomes Creek
#> 25 Thomes Creek
#> 26 Thomes Creek
#> 27 Thomes Creek
#> 28 Thomes Creek
#> 29 Thomes Creek
#> 30 Thomes Creek
#> 31 Thomes Creek
#> 32 Thomes Creek