Tuolumne River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

tuolumne_river_floodplain

Format

dataframe with 37 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

TUFLOW hydraulic model with 1D channel and 2D overbank components

Fall Run, Spring Run and Steelhead Modeling

The entire mapped rearing extent of 54.33 miles was modeled by TID and MID using a TUFLOW hydraulic model with 1D channel and 2D overbank components. This approach directly modeled inundated floodplain area.

Examples

tuolumne_river_floodplain

#> # A tibble: 37 x 5
#>    flow_cfs FR_floodplain_acr… SR_floodplain_acr… ST_floodplain_ac… watershed  
#>       <dbl>              <dbl>              <dbl>             <dbl> <chr>      
#>  1      100                0                  0                 0   Tuolumne R…
#>  2      250                0                  0                 0   Tuolumne R…
#>  3      500                0                  0                 0   Tuolumne R…
#>  4      750                0                  0                 0   Tuolumne R…
#>  5     1000                0                  0                 0   Tuolumne R…
#>  6     1250               36.0               36.0              36.0 Tuolumne R…
#>  7     1500               76.7               76.7              76.7 Tuolumne R…
#>  8     1750              117.               117.              117.  Tuolumne R…
#>  9     2000              157.               157.              157.  Tuolumne R…
#> 10     2250              197.               197.              197.  Tuolumne R…
#> # ... with 27 more rows