Mainstem Sacramento River Spawning Habitat

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second.

upper_sac_ACID_boards_in

upper_sac_ACID_boards_out

Format

dataframe with 30 rows and 11 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_WUA

fall run spawning WUA in square feet per 1000 feet

LFR_spawn_WUA

late-fall run fry spawning WUA in square feet per 1000 feet

ST_spawn_WUA

steelhead spawning WUA in square feet per 1000 feet

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

The spawning WUA were derived from a FWS River2D model. The study is broken into several segments: Keswick to A.C.I.D. (3.5 mi), A.C.I.D. to Cow Creek (18.5 mi), Cow Creek to Battle Creek (8.5 mi), Battle Creek to Red Bluff (22.5 mi), and Red Bluff to Deer Creek (23.5 mi). The Upper Sacramento fall run spawning WUA values include the spawning that occurs in the Upper and Upper-mid Sacramento River (Keswick to Deer Creek). The Late-Fall Run, Winter Run, and Steelhead spawn from Keswick to Battle Creek. The A.C.I.D. boards are in April 1st - October 31st

Species			Migration

Peak Migration			Spawning

Peak Spawning			Juvenile Emergence			Juvenile Rearing

Late-Fall Run	Oct-Apr	Dec	Jan-Apr	Feb-Mar

Apr-Jun	7-13 mths		Winter Run	Dec-Jul	Mar	Apr-Aug

May-Jun	Jul-Oct	5-10 mths		Spring Run	Mar-Sep	May-Jun

Aug-Oct	Mid-Sep	Nov-Mar	3-15 mths		Fall Run	Jun-Dec

Sep-Oct	Sep-Dec	Oct-Nov	Dec-Mar	1-7 mths

Generalised life history Yoshinyama et al. 1998

Examples

upper_sac_ACID_boards_in

#> # A tibble: 30 x 6
#>    flow_cfs FR_spawn_WUA LFR_spawn_WUA ST_spawn_WUA WR_spawn_WUA watershed     
#>       <dbl>        <dbl>         <dbl>        <dbl>        <dbl> <chr>         
#>  1     3250        8554.         7945.        1998.        5321. Upper Sacrame…
#>  2     3500        8757.         8204.        2035.        5627. Upper Sacrame…
#>  3     3750        8940.         8330.        2041.        5881. Upper Sacrame…
#>  4     4000        9069.         8395.        2025.        6111. Upper Sacrame…
#>  5     4250        9082.         8429.        2009.        6350. Upper Sacrame…
#>  6     4500        9193.         8423.        2005.        6580. Upper Sacrame…
#>  7     4750        9312.         8375.        1999.        6796. Upper Sacrame…
#>  8     5000        9282.         8310.        1997.        6982. Upper Sacrame…
#>  9     5250        9299.         8229.        2003.        7126. Upper Sacrame…
#> 10     5500        9254.         8120.        2008.        7278. Upper Sacrame…
#> # ... with 20 more rows