Upper-Mid Sacramento River Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

upper_mid_sacramento_river_instream

Format

dataframe with 45 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

The rearing WUA was derived from a HEC-RAS 1D model. The study's extent is from Battle Creek to the confluence with the Feather River (189.1 mi). The CVPIA Upper-mid Sacramento River extends from Red Blurr to Wilkins Slough (122.45 mi).

Examples

upper_mid_sacramento_river_instream

#> # A tibble: 45 x 3
#>    flow_cfs FR_juv_wua                  watershed
#>       <dbl>      <dbl>                      <chr>
#>  1 2961.333  10000.450 Upper-mid Sacramento River
#>  2 3109.146   9725.913 Upper-mid Sacramento River
#>  3 3384.049   8785.687 Upper-mid Sacramento River
#>  4 3596.065   8779.428 Upper-mid Sacramento River
#>  5 3898.528   8439.424 Upper-mid Sacramento River
#>  6 4144.839   8130.594 Upper-mid Sacramento River
#>  7 4388.373   7485.262 Upper-mid Sacramento River
#>  8 4619.236   7106.064 Upper-mid Sacramento River
#>  9 4853.909   7921.588 Upper-mid Sacramento River
#> 10 5129.724   7682.911 Upper-mid Sacramento River
#> # ... with 35 more rows