Upper Sacramento River Instream Flow to Habitat Area Relationship (A.C.I.D. Boards In)

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second.

upper_sac_ACID_boards_in

Format

dataframe with 30 rows and 11 variables:

flow_cfs

integer flow value in cubic feet per second

FR_fry_WUA

fall run fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_WUA

fall run juvenile WUA in square feet per 1000 feet

FR_spawn_WUA

fall run spawning WUA in square feet per 1000 feet

LFR_fry_WUA

late-fall run fry (up to 50 mm) WUA in square feet per 1000 feet

LFR_juv_WUA

late-fall run fry juvenile WUA in square feet per 1000 feet

LFR_spawn_WUA

late-fall run fry spawning WUA in square feet per 1000 feet

ST_spawn_WUA

steelhead spawning WUA in square feet per 1000 feet

WR_fry_WUA

winter run fry (up to 50 mm) WUA in square feet per 1000 feet

WR_juv_WUA

winter run juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

FWS and NMFS Life Cycle Model, Hendrix 2017

Details

The spawning WUA were derived from a FWS River2D model. The study is broken into several segments: Keswick to A.C.I.D. (3.5 mi), A.C.I.D. to Cow Creek (18.5 mi), Cow Creek to Battle Creek (8.5 mi), Battle Creek to Red Bluff (22.5 mi), and Red Bluff to Deer Creek (23.5 mi). The Upper Sacramento fall run spawning WUA values include the spawning that occurs in the Upper and Upper-mid Sacramento River (Keswick to Deer Creek). The Late-Fall Run, Winter Run, and Steelhead spawn from Keswick to Battle Creek. The rearing WUA values come from two models, River2D (Keswick to Battle Creek) and HEC-RAS 1D (Battle Creek to Red Bluff). The A.C.I.D. boards are in April 1st - October 31st

Species 	 Migration 	 Peak Migration 	 Spawning 	 Peak Spawning 	 Juvenile Emergence 	 Juvenile Rearing

Late-Fall Run 	 Oct-Apr 	 Dec 	 Jan-Apr 	 Feb-Mar 	 Apr-Jun 	 7-13 mths

Winter Run 	 Dec-Jul 	 Mar 	 Apr-Aug 	 May-Jun 	 Jul-Oct 	 5-10 mths

Spring Run 	 Mar-Sep 	 May-Jun 	 Aug-Oct 	 Mid-Sep 	 Nov-Mar 	 3-15 mths

Fall Run 	 Jun-Dec 	 Sep-Oct 	 Sep-Dec 	 Oct-Nov 	 Dec-Mar 	 1-7 mths

Generalised life history Yoshinyama et al. 1998

Use these values to compute spawning habitat for Winter Run. Use these values to compute rearing habitat for ....