Upper Sacramento River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

upper_sacramento_river_floodplain

Format

dataframe with 240 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

The habitat area was derived from a HEC-RAS 1D model, Hendrix 2017. The CVPIA Upper Sacramento River extends from Keswick to Red Bluff (59.2 mi). This reach overlaps with two of the study's, Keswick to Battle Creek (55.5 mi) and Battle Creek to the confluence with the Feather River (189.1 mi). To scale the study's results to CVPIA's extents, the floodplain acerage at a given flow is the weighted average of the floodplain area per river mile within each study.

Fall Run and Steelhead Modeling

The entire mapped rearing extent of 59.19 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model. The high quality depth and high quality velocity ("Pref11") "BankArea" result was used as floodplain area. High quality velocities were assumed to be less than or equal to 0.15 meters per second, and high quality depths were assumed to be between 0.2 meters and 1.5 meters.

Examples

upper_sacramento_river_floodplain

#> # A tibble: 240 x 5
#>    flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#>       <dbl>               <dbl>               <dbl>               <dbl>
#>  1 2637.426            12.96263            12.96263            12.96263
#>  2 2967.296            17.39277            17.39277            17.39277
#>  3 2992.467            17.43650            17.43650            17.43650
#>  4 3384.049            17.93253            17.93253            17.93253
#>  5 3479.393            18.09580            18.09580            18.09580
#>  6 3550.317            18.22282            18.22282            18.22282
#>  7 3854.895            18.73746            18.73746            18.73746
#>  8 3897.321            18.78137            18.78137            18.78137
#>  9 4144.839            22.49003            22.49003            22.49003
#> 10 4355.361            25.72022            25.72022            25.72022
#> # ... with 230 more rows, and 1 more variables: watershed <chr>