Flow Thresholds for Number of Weeks Inundated

Estimated relationship between duration of inundation and monthly mean flow

weeks_inundated

Format

dataframe with rows and 3 variables

watershed

name of CVPIA watershed

weeks_inundated

integer value between 0 and 4 to represent weeks of floodplain inundation

flow_threshold

flow threshold in cubic feet per second associated with number of weeks inundated

Source

Sadie Gill sgill@flowwest.com

Details

Visual inspection of measured flow and professional judgement were used to define relationships between number of days inundated and the mean monthly flow.

Relationship Modeled:

Sacramento River

Big Chico Creek

Butte Creek

Cottonwood Creek

Deer Creek

Elder Creek

Sutter Bypass

Bear River

Feather River

Yuba River

Yolo Bypass

American River

Calaveras River

Merced River

Stanislaus River

Tuolumne River

San Joaquin River

All other watersheds are assumed to have a two week inundation.

Examples

weeks_inundated

#>          watershed weeks_inundated flow_threshhold
#> 1       Yuba River               0               0
#> 2       Yuba River               1               0
#> 3       Yuba River               2               0
#> 4       Yuba River               3               0
#> 5       Yuba River               4             880
#> 6  Big Chico Creek               0               0
#> 7  Big Chico Creek               1               0
#> 8  Big Chico Creek               2             265
#> 9  Big Chico Creek               3               0
#> 10 Big Chico Creek               4             900