WUA to Area in Square Meters

Function converts a Weighted Usable Area (WUA) to an area in square meters by multiplying the WUA by the stream's length used by the species. The lengths were obtained through expert outreach in 2017 and are available here watershed_lengths.

wua_to_area(wua, watershed_name, life_stage, species_name)

Arguments

wua
WUA in square feet per 1000 feet

life_stage
life stage

watershed
watershed

species
species, "fr" - fall run or "sr" - spring run or "st" - steel head