Yuba River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

yuba_river_floodplain

Format

dataframe with 20 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage

watershed

name of watershed

Source

SRH2D

Details

The habitat area was derived from the SRH2D model, Pasternack 2012.

Examples

yuba_river_floodplain

#> # A tibble: 20 x 3
#>    flow_cfs floodplain_acres  watershed
#>       <int>            <dbl>      <chr>
#>  1      800           0.0000 Yuba River
#>  2      880           0.9942 Yuba River
#>  3      930           1.6135 Yuba River
#>  4     1000           2.4373 Yuba River
#>  5     1300           5.6163 Yuba River
#>  6     1500          27.1855 Yuba River
#>  7     1700          42.2835 Yuba River
#>  8     2000          65.6656 Yuba River
#>  9     2500         100.3747 Yuba River
#> 10     3000         137.4171 Yuba River
#> 11     4000         206.1480 Yuba River
#> 12     5000         266.6243 Yuba River
#> 13     7500         406.6790 Yuba River
#> 14    10000         555.8941 Yuba River
#> 15    15000         875.1997 Yuba River
#> 16    21100        1132.7503 Yuba River
#> 17    30000        1389.4573 Yuba River
#> 18    42200        1574.8585 Yuba River
#> 19    84400        2201.1325 Yuba River
#> 20   110400        2561.9083 Yuba River