Battle Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

battle_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Battle Creek scaling factor of 147% was calculated as the average ratio of Battle Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 5.87 mile low gradient mapped rearing extent in Battle Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 52.4 acres) and floodplain inundation area.

Examples

battle_creek_floodplain

#>     flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1   103.0865            0.000000            0.000000            0.000000
#> 2   250.3529            0.000000            0.000000            0.000000
#> 3   397.6193            0.000000            0.000000            0.000000
#> 4   544.8858            0.000000            0.000000            0.000000
#> 5   692.1522            0.000000            0.000000            0.000000
#> 6   839.4186            0.000000            0.000000            0.000000
#> 7   840.8913            0.000000            0.000000            0.000000
#> 8   986.6850            0.000000            0.000000            0.000000
#> 9  1133.9515            0.000000            0.000000            0.000000
#> 10 1281.2179            1.808845            2.380058            2.380058
#> 11 1428.4843            5.312169            6.989692            6.989692
#> 12 1575.7507            8.250600           10.856046           10.856046
#> 13 1723.0172           11.410860           15.014280           15.014280
#> 14 1870.2836           14.205055           18.690850           18.690850
#> 15 2017.5500           17.184476           22.611138           22.611138
#> 16 2164.8164           19.997166           26.312044           26.312044
#> 17 2312.0828           22.718087           29.892201           29.892201
#> 18 2459.3493           25.053721           32.965401           32.965401
#> 19 2606.6157           27.289401           35.907083           35.907083
#> 20 2753.8821           29.392388           38.674170           38.674170
#> 21 2901.1485           31.248856           41.116889           41.116889
#> 22 3048.4150           33.204097           43.689573           43.689573
#> 23 3195.6814           34.727508           45.694060           45.694060
#> 24 3342.9478           36.260589           47.711271           47.711271
#> 25 3490.2142           37.589002           49.459182           49.459182
#> 26 3637.4807           39.042819           51.372097           51.372097
#> 27 3784.7471           40.447679           53.220596           53.220596
#> 28 3932.0135           41.821505           55.028261           55.028261
#> 29 4079.2799           43.132638           56.753435           56.753435
#> 30 4226.5463           44.186046           58.139498           58.139498
#>       watershed
#> 1  Battle Creek
#> 2  Battle Creek
#> 3  Battle Creek
#> 4  Battle Creek
#> 5  Battle Creek
#> 6  Battle Creek
#> 7  Battle Creek
#> 8  Battle Creek
#> 9  Battle Creek
#> 10 Battle Creek
#> 11 Battle Creek
#> 12 Battle Creek
#> 13 Battle Creek
#> 14 Battle Creek
#> 15 Battle Creek
#> 16 Battle Creek
#> 17 Battle Creek
#> 18 Battle Creek
#> 19 Battle Creek
#> 20 Battle Creek
#> 21 Battle Creek
#> 22 Battle Creek
#> 23 Battle Creek
#> 24 Battle Creek
#> 25 Battle Creek
#> 26 Battle Creek
#> 27 Battle Creek
#> 28 Battle Creek
#> 29 Battle Creek
#> 30 Battle Creek