Bear Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

bear_creek_floodplain

Format

dataframe with 30 rows and 4 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Fall Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Deer Creek based onhydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.25 percent of DeerCreek values based on the ratio of mean flow from December to June between this watershed and Deer Creek. Of the entire mapped 17.30 mile rearing extent in this watershed, 4.32 miles were classified as low gradient and 12.98 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

bear_creek_floodplain

#>     flow_cfs FR_floodplain_acres ST_floodplain_acres  watershed
#> 1   17.77286           0.0000000           0.0000000 Bear Creek
#> 2   43.16265           0.0000000           0.0000000 Bear Creek
#> 3   68.55245           0.0000000           0.0000000 Bear Creek
#> 4   93.94224           0.0000000           0.0000000 Bear Creek
#> 5  119.33204           0.0000000           0.0000000 Bear Creek
#> 6  144.72183           0.0000000           0.0000000 Bear Creek
#> 7  144.97573           0.0000000           0.0000000 Bear Creek
#> 8  170.11163           0.0000000           0.0000000 Bear Creek
#> 9  195.50142           0.0000000           0.0000000 Bear Creek
#> 10 220.89122           0.2984411           0.2984411 Bear Creek
#> 11 246.28101           0.8764541           0.8764541 Bear Creek
#> 12 271.67081           1.3612654           1.3612654 Bear Creek
#> 13 297.06060           1.8826763           1.8826763 Bear Creek
#> 14 322.45040           2.3436901           2.3436901 Bear Creek
#> 15 347.84019           2.8352644           2.8352644 Bear Creek
#> 16 373.22999           3.2993298           3.2993298 Bear Creek
#> 17 398.61978           3.7482541           3.7482541 Bear Creek
#> 18 424.00958           4.1336101           4.1336101 Bear Creek
#> 19 449.39937           4.5024746           4.5024746 Bear Creek
#> 20 474.78917           4.8494461           4.8494461 Bear Creek
#> 21 500.17896           5.1557445           5.1557445 Bear Creek
#> 22 525.56876           5.4783394           5.4783394 Bear Creek
#> 23 550.95855           5.7296868           5.7296868 Bear Creek
#> 24 576.34835           5.9826297           5.9826297 Bear Creek
#> 25 601.73814           6.2018044           6.2018044 Bear Creek
#> 26 627.12794           6.4416694           6.4416694 Bear Creek
#> 27 652.51773           6.6734571           6.6734571 Bear Creek
#> 28 677.90753           6.9001244           6.9001244 Bear Creek
#> 29 703.29732           7.1164481           7.1164481 Bear Creek
#> 30 728.68712           7.2902499           7.2902499 Bear Creek