Butte Creek Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

butte_creek_instream

Format

dataframe with 41 rows and 6 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

adult_trout_WUA

adult trout WUA in square feet per 1000 feet

watershed

name of watershed

Source

FWS and 2008 FERC relicensing of DeSabla

Details

The spawning WUA values were derived from a FWS River2D model and the rearing WUA from a PHABSIM model created for the 2008 FERC relicensing of DeSabla. The spawning values are from two river segments, above Centerville Powerhouse (6.5 miles) and below (9 miles). The spawning results are for Spring Run Chinook. The rearing results are for Fall Run Chinook.

Examples

butte_creek_instream

#> # A tibble: 41 x 6
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua adult_trout_WUA watershed  
#>       <int>        <dbl>      <int>      <int>           <int> <chr>      
#>  1       20        1074.         NA         NA              NA Butte Creek
#>  2       25        1186.         NA         NA              NA Butte Creek
#>  3       30        1291.         NA         NA              NA Butte Creek
#>  4       35        1370.         NA         NA              NA Butte Creek
#>  5       40        1450.         NA         NA              NA Butte Creek
#>  6       45        1528.         NA         NA              NA Butte Creek
#>  7       50        1584.         NA         NA              NA Butte Creek
#>  8       60        1645.      17549      18472           14399 Butte Creek
#>  9       70        1690.      16339      19560           15950 Butte Creek
#> 10       80        1721.      15208      20259           17296 Butte Creek
#> # ... with 31 more rows