Clear Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second

clear_creek_floodplain

Format

dataframe with 32 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled USFWS / FEMA 1D HEC-RAS hydraulic model

Fall Run Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.21 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 8.00 mile rearing extent in this watershed, 8.00 miles were classified as low gradient and -0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Spring Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.21 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 17.92 mile rearing extent in this watershed, 8.00 miles were classified as low gradient and 9.92 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

clear_creek_floodplain

#>       flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1     8.369479           0.0000000           0.0000000           0.0000000
#> 2    83.694785           0.0000000           0.0000000           0.0000000
#> 3   159.020092           0.0000000           0.0000000           0.0000000
#> 4   234.345399           0.0000000           0.0000000           0.0000000
#> 5   309.670705           0.0000000           0.0000000           0.0000000
#> 6   384.996012           0.0000000           0.0000000           0.0000000
#> 7   391.273121           0.0000000           0.0000000           0.0000000
#> 8   460.321319           0.0000000           0.0000000           0.0000000
#> 9   535.646626           0.0000000           0.0000000           0.0000000
#> 10  610.971932           0.0000000           0.0000000           0.0000000
#> 11  686.297239           0.0000000           0.0000000           0.0000000
#> 12  761.622546           0.0000000           0.0000000           0.0000000
#> 13  836.947853           0.0000000           0.0000000           0.0000000
#> 14  912.273159           0.0000000           0.0000000           0.0000000
#> 15  987.598466           0.0000000           0.0000000           0.0000000
#> 16 1062.923773           0.0000000           0.0000000           0.0000000
#> 17 1138.249080           0.0000000           0.0000000           0.0000000
#> 18 1213.574386           0.0000000           0.0000000           0.0000000
#> 19 1288.899693           0.0000000           0.0000000           0.0000000
#> 20 1364.225000           0.0000000           0.0000000           0.0000000
#> 21 1439.550307           0.0000000           0.0000000           0.0000000
#> 22 1514.875613           0.0000000           0.0000000           0.0000000
#> 23 1590.200920           0.0000000           0.0000000           0.0000000
#> 24 1665.526227           0.4424499           0.4973432           0.4973432
#> 25 1740.851533           0.9927313           1.1158963           1.1158963
#> 26 1816.176840           1.4421274           1.6210475           1.6210475
#> 27 1891.502147           1.8118601           2.0366519           2.0366519
#> 28 1966.827454           2.1844439           2.4554609           2.4554609
#> 29 2042.152760           2.4998287           2.8099745           2.8099745
#> 30 2117.478067           2.8470696           3.2002965           3.2002965
#> 31 2192.803374           3.2320455           3.6330352           3.6330352
#> 32 2268.128681           3.5571004           3.9984186           3.9984186
#>      watershed
#> 1  Clear Creek
#> 2  Clear Creek
#> 3  Clear Creek
#> 4  Clear Creek
#> 5  Clear Creek
#> 6  Clear Creek
#> 7  Clear Creek
#> 8  Clear Creek
#> 9  Clear Creek
#> 10 Clear Creek
#> 11 Clear Creek
#> 12 Clear Creek
#> 13 Clear Creek
#> 14 Clear Creek
#> 15 Clear Creek
#> 16 Clear Creek
#> 17 Clear Creek
#> 18 Clear Creek
#> 19 Clear Creek
#> 20 Clear Creek
#> 21 Clear Creek
#> 22 Clear Creek
#> 23 Clear Creek
#> 24 Clear Creek
#> 25 Clear Creek
#> 26 Clear Creek
#> 27 Clear Creek
#> 28 Clear Creek
#> 29 Clear Creek
#> 30 Clear Creek
#> 31 Clear Creek
#> 32 Clear Creek