Clear Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second

clear_creek_floodplain

Format

dataframe with 8 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

NA

Details

HEC-RAS model runs to generate Cottonwood Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Clear Creek scaling factor of 21% was calculated as the average ratio of Clear Creek monthly mean flow to Cottonwood Creek monthly mean flow between December and June. This scaling factor was applied to Cottonwood Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 8 mile mapped rearing extent in Clear Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 57.4 acres) and floodplain inundation area.

Examples

clear_creek_floodplain

#> # A tibble: 8 x 3
#>   flow_cfs floodplain_acres watershed  
#>      <dbl>            <dbl> <chr>      
#> 1     1590            0     Clear Creek
#> 2     1666            0.440 Clear Creek
#> 3     1741            0.990 Clear Creek
#> 4     1816            1.44  Clear Creek
#> 5     1892            1.81  Clear Creek
#> 6     1967            2.18  Clear Creek
#> 7     2042            2.50  Clear Creek
#> 8     2117            2.84  Clear Creek