Cosumnes River Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

cosumnes_river_floodplain

Format

dataframe with 36 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage (use for Steelhead when needed)

watershed

name of watershed

Source

NA

Details

NA

Examples

cosumnes_river_floodplain

#> # A tibble: 35 x 3
#>    flow_cfs FR_floodplain_acres      watershed
#>       <int>               <dbl>          <chr>
#>  1       10                  NA Cosumnes River
#>  2      210                  NA Cosumnes River
#>  3      410                  NA Cosumnes River
#>  4      610                  NA Cosumnes River
#>  5      810                  NA Cosumnes River
#>  6     1010                  NA Cosumnes River
#>  7     1122               106.1 Cosumnes River
#>  8     1210                  NA Cosumnes River
#>  9     1410                  NA Cosumnes River
#> 10     1610                  NA Cosumnes River
#> # ... with 25 more rows