Cosumnes River Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

cosumnes_river_instream

Format

dataframe with 14 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Sadie Gill

Details

The Cosumnes River instream rearing habitat has not been modeled. The WUA values were estimated using the mean WUA at each flow from Mokelumne and Calaveras Rivers.

Examples

cosumnes_river_instream

#> # A tibble: 10 x 5
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua watershed     
#>       <dbl>        <dbl>      <dbl>      <dbl> <chr>         
#>  1      100        3446.     14392.     16453. Cosumnes River
#>  2      200        4624.     12140.     14936. Cosumnes River
#>  3      300        4996.     10588.     14390. Cosumnes River
#>  4      400        5717.      9412.     12934. Cosumnes River
#>  5      500        6134.      8129.     11238. Cosumnes River
#>  6      600        6134.      7537.     10278. Cosumnes River
#>  7      700        6134.      7128.      9291. Cosumnes River
#>  8      800        6134.      6759.      8913. Cosumnes River
#>  9      900        6134.      6428.      8359. Cosumnes River
#> 10     1000        6134.      6214.      8119. Cosumnes River