Cow Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

cow_creek_floodplain

Format

dataframe with 30 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Fall Run Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Deer Creek based onhydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 2.16 percent of DeerCreek values based on the ratio of mean flow from December to June between this watershed and Deer Creek. Of the entire mapped 64.17 mile rearing extent in this watershed, 8.59 miles were classified as low gradient and 55.58 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Deer Creek based onhydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 2.16 percent of DeerCreek values based on the ratio of mean flow from December to June between this watershed and Deer Creek. Of the entire mapped 94.19 mile rearing extent in this watershed, 8.59 miles were classified as low gradient and 85.60 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

cow_creek_floodplain

#>     flow_cfs FR_floodplain_acres ST_floodplain_acres watershed
#> 1   150.9354            0.000000            0.000000 Cow Creek
#> 2   366.5575            0.000000            0.000000 Cow Creek
#> 3   582.1796            0.000000            0.000000 Cow Creek
#> 4   797.8017            0.000000            0.000000 Cow Creek
#> 5  1013.4237            0.000000            0.000000 Cow Creek
#> 6  1229.0458            0.000000            0.000000 Cow Creek
#> 7  1231.2020            0.000000            0.000000 Cow Creek
#> 8  1444.6679            0.000000            0.000000 Cow Creek
#> 9  1660.2899            0.000000            0.000000 Cow Creek
#> 10 1875.9120            6.383165            7.737547 Cow Creek
#> 11 2091.5341           18.745913           22.723428 Cow Creek
#> 12 2307.1562           29.115228           35.292909 Cow Creek
#> 13 2522.7782           40.267350           48.811293 Cow Creek
#> 14 2738.4003           50.127678           60.763789 Cow Creek
#> 15 2954.0224           60.641644           73.508613 Cow Creek
#> 16 3169.6444           70.567240           85.540225 Cow Creek
#> 17 3385.2665           80.168994           97.179283 Cow Creek
#> 18 3600.8886           88.411124          107.170232 Cow Creek
#> 19 3816.5107           96.300529          116.733614 Cow Creek
#> 20 4032.1327          103.721679          125.729387 Cow Creek
#> 21 4247.7548          110.272899          133.670647 Cow Creek
#> 22 4463.3769          117.172674          142.034420 Cow Creek
#> 23 4678.9989          122.548582          148.550990 Cow Creek
#> 24 4894.6210          127.958613          155.108924 Cow Creek
#> 25 5110.2431          132.646402          160.791369 Cow Creek
#> 26 5325.8652          137.776719          167.010238 Cow Creek
#> 27 5541.4872          142.734276          173.019692 Cow Creek
#> 28 5757.1093          147.582318          178.896394 Cow Creek
#> 29 5972.7314          152.209125          184.504919 Cow Creek
#> 30 6188.3534          155.926460          189.010999 Cow Creek