Deer Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

deer_creek_floodplain

Format

dataframe with 30 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage (use for Steelhead when needed)

watershed

name of watershed

Source

TODO

Details

The habitat area was derived from the Deer Creek Watershed Conservancy HEC-RAS 1D 2011 model.

Examples

deer_creek_floodplain

#> # A tibble: 30 x 3
#>    flow_cfs FR_floodplain_acres  watershed
#>       <int>               <dbl>      <chr>
#>  1       70           0.0000000 Deer Creek
#>  2      170           0.0000000 Deer Creek
#>  3      270           0.0000000 Deer Creek
#>  4      370           0.0000000 Deer Creek
#>  5      470           0.4015106 Deer Creek
#>  6      570           6.2421212 Deer Creek
#>  7      571           6.2986260 Deer Creek
#>  8      670          11.8926033 Deer Creek
#>  9      770          18.2723783 Deer Creek
#> 10      870          23.4906749 Deer Creek
#> # ... with 20 more rows