Elder Creek Flow to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

elder_creek_floodplain

Format

dataframe with 31 rows and 4 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model

Fall Run and Steelhead Modeling

A 5.40 mile portion of the entire mapped rearing extent of 7.10 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 7.10 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

elder_creek_floodplain

#>    flow_cfs FR_floodplain_acres ST_floodplain_acres   watershed
#> 1        10            0.000000            0.000000 Elder Creek
#> 2        50            0.000000            0.000000 Elder Creek
#> 3        90            0.000000            0.000000 Elder Creek
#> 4       130            5.115882            5.115882 Elder Creek
#> 5       170           13.912115           13.912115 Elder Creek
#> 6       196           16.602674           16.602674 Elder Creek
#> 7       210           18.051436           18.051436 Elder Creek
#> 8       250           21.356318           21.356318 Elder Creek
#> 9       290           24.463110           24.463110 Elder Creek
#> 10      330           27.105953           27.105953 Elder Creek
#> 11      370           29.462529           29.462529 Elder Creek
#> 12      410           31.227916           31.227916 Elder Creek
#> 13      450           32.860780           32.860780 Elder Creek
#> 14      490           34.447035           34.447035 Elder Creek
#> 15      530           35.778358           35.778358 Elder Creek
#> 16      570           37.057487           37.057487 Elder Creek
#> 17      610           38.333899           38.333899 Elder Creek
#> 18      650           39.491765           39.491765 Elder Creek
#> 19      690           40.634296           40.634296 Elder Creek
#> 20      730           41.751410           41.751410 Elder Creek
#> 21      770           42.763713           42.763713 Elder Creek
#> 22      810           43.923964           43.923964 Elder Creek
#> 23      850           44.829917           44.829917 Elder Creek
#> 24      890           45.665593           45.665593 Elder Creek
#> 25      930           46.468486           46.468486 Elder Creek
#> 26      970           47.389231           47.389231 Elder Creek
#> 27     1010           48.145183           48.145183 Elder Creek
#> 28     1050           48.967275           48.967275 Elder Creek
#> 29     1090           49.635563           49.635563 Elder Creek
#> 30     1130           50.389794           50.389794 Elder Creek
#> 31     1170           51.063666           51.063666 Elder Creek