Mill Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

mill_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Mill Creek scaling factor of 98

Examples

mill_creek_floodplain

#>      flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1    68.72368            0.000000            0.000000            0.000000
#> 2   166.90036            0.000000            0.000000            0.000000
#> 3   265.07705            0.000000            0.000000            0.000000
#> 4   363.25373            0.000000            0.000000            0.000000
#> 5   461.43042            0.000000            0.000000            0.000000
#> 6   559.60710            0.000000            0.000000            0.000000
#> 7   560.58887            0.000000            0.000000            0.000000
#> 8   657.78379            0.000000            0.000000            0.000000
#> 9   755.96047            0.000000            0.000000            0.000000
#> 10  854.13716            1.820055            2.485458            2.485458
#> 11  952.31384            5.345091            7.299229            7.299229
#> 12 1050.49053            8.301732           11.336802           11.336802
#> 13 1148.66721           11.481577           15.679183           15.679183
#> 14 1246.84390           14.293089           19.518568           19.518568
#> 15 1345.02058           17.290974           23.612466           23.612466
#> 16 1443.19726           20.121096           27.477265           27.477265
#> 17 1541.37395           22.858879           31.215968           31.215968
#> 18 1639.55063           25.208988           34.425264           34.425264
#> 19 1737.72732           27.458523           37.497218           37.497218
#> 20 1835.90400           29.574543           40.386844           40.386844
#> 21 1934.08069           31.442516           42.937739           42.937739
#> 22 2032.25737           33.409875           45.624353           45.624353
#> 23 2130.43406           34.942727           47.717608           47.717608
#> 24 2228.61074           36.485309           49.824150           49.824150
#> 25 2326.78743           37.821955           51.649468           51.649468
#> 26 2424.96411           39.284781           53.647095           53.647095
#> 27 2523.14080           40.698348           55.577455           55.577455
#> 28 2621.31748           42.080688           57.465171           57.465171
#> 29 2719.49417           43.399946           59.266744           59.266744
#> 30 2817.67085           44.459883           60.714188           60.714188
#>     watershed
#> 1  Mill Creek
#> 2  Mill Creek
#> 3  Mill Creek
#> 4  Mill Creek
#> 5  Mill Creek
#> 6  Mill Creek
#> 7  Mill Creek
#> 8  Mill Creek
#> 9  Mill Creek
#> 10 Mill Creek
#> 11 Mill Creek
#> 12 Mill Creek
#> 13 Mill Creek
#> 14 Mill Creek
#> 15 Mill Creek
#> 16 Mill Creek
#> 17 Mill Creek
#> 18 Mill Creek
#> 19 Mill Creek
#> 20 Mill Creek
#> 21 Mill Creek
#> 22 Mill Creek
#> 23 Mill Creek
#> 24 Mill Creek
#> 25 Mill Creek
#> 26 Mill Creek
#> 27 Mill Creek
#> 28 Mill Creek
#> 29 Mill Creek
#> 30 Mill Creek