Habitat Modeling Status

A table lifestage, species, and watershed

modeling_exist

Format

dataframe with 33 rows and 15 variables

Order

integer value representing watershed order in SIT model

Watershed

name of watershed

FR_spawn

TRUE if spawning habitat modeling exists for Fall Run

FR_fry

TRUE if fry rearing habitat modeling exists for Fall Run

FR_juv

TRUE if juvenile rearing habitat modeling exists for Fall Run

FR_floodplain

TRUE if floodplain rearing modeling exists for Fall Run

SR_spawn

TRUE if spawning habitat modeling exists for Spring Run

SR_fry

TRUE if fry rearing habitat modeling exists for Spring Run

SR_juv

TRUE if juvenile rearing habitat modeling exists for Spring Run

SR_floodplain

TRUE if floodplain rearing modeling exists for Spring Run

ST_spawn

TRUE if spawning habitat modeling exists for Steelhead

ST_fry

TRUE if fry rearing habitat modeling exists for Steelhead

ST_juv

TRUE if juvenile rearing habitat modeling exists for Steelhead

ST_floodplain

TRUE if floodplain rearing modeling exists for Steelhead

Region

Regional grouping of watersheds used for estimating values for watersheds without modeling

UseRearRegionApprox

boolean flag for set_habitat functions, TRUE denotes no rearing modeling exists for watershed

UseSpawnRegionApprox

boolean flag for set_habitat functions, TRUE denotes no spawning modeling exists for watershed

UseFloodplainRegionApprox

boolean flag for set_habitat functions, TRUE denotes no floodplain rearing modeling exists for watershed

Source

Sadie Gill sgill@flowwest.com

Details

This table was compiled from information provided by Mark Gard mark_gard@fws.gov. Used as look up table for set_spawning_habitat, set_instream_habitat, and set_floodplain_habitat

Examples

modeling_exist

#> # A tibble: 32 x 18
#>    Order Watershed FR_spawn FR_fry FR_juv FR_floodplain SR_spawn SR_fry SR_juv
#>    <int> <chr>     <lgl>    <lgl>  <lgl>  <lgl>         <lgl>    <lgl>  <lgl> 
#>  1     1 Upper Sa… TRUE     FALSE  TRUE   TRUE          FALSE    FALSE  FALSE 
#>  2     2 Antelope… FALSE    FALSE  FALSE  FALSE         FALSE    FALSE  FALSE 
#>  3     3 Battle C… TRUE     TRUE   TRUE   FALSE         FALSE    FALSE  FALSE 
#>  4     4 Bear Cre… FALSE    FALSE  FALSE  FALSE         NA       NA     NA    
#>  5     5 Big Chic… FALSE    FALSE  FALSE  TRUE          FALSE    FALSE  FALSE 
#>  6     6 Butte Cr… TRUE     TRUE   TRUE   TRUE          FALSE    FALSE  FALSE 
#>  7     7 Clear Cr… TRUE     TRUE   TRUE   FALSE         TRUE     TRUE   TRUE  
#>  8     8 Cottonwo… TRUE     TRUE   TRUE   TRUE          FALSE    FALSE  FALSE 
#>  9     9 Cow Creek FALSE    TRUE   TRUE   FALSE         NA       NA     NA    
#> 10    10 Deer Cre… FALSE    FALSE  FALSE  TRUE          FALSE    FALSE  FALSE 
#> # ... with 22 more rows, and 9 more variables: SR_floodplain <lgl>,
#> #   ST_spawn <lgl>, ST_fry <lgl>, ST_juv <lgl>, ST_floodplain <lgl>,
#> #   Region <chr>, UseRearRegionApprox <lgl>, UseSpawnRegionApprox <lgl>,
#> #   UseFloodplainRegionApprox <lgl>