Mokelumne River to Floodplain Habitat Area Relationship

A dataset containing the floodplain habitat area in acres as a function of flow in cubic feet per second

mokelumne_river_floodplain

Format

dataframe with 37 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

SR_floodplain_acres

spring run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

TUFLOW hydraulic model with 1D channel and 2D overbank components

Fall Run, Spring Run and Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for the Tuolumne River based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.50 percent of the Tuolumne River values based on the ratio of mean flow from December to June between this watershed and the Tuolumne River. Of the entire mapped 58.40 mile rearing extent in this watershed, 58.40 miles were classifiedas low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

mokelumne_river_floodplain

#>     flow_cfs FR_floodplain_acres SR_floodplain_acres ST_floodplain_acres
#> 1    50.2672             0.00000             0.00000             0.00000
#> 2   125.6680             0.00000             0.00000             0.00000
#> 3   251.3360             0.00000             0.00000             0.00000
#> 4   377.0040             0.00000             0.00000             0.00000
#> 5   502.6720             0.00000             0.00000             0.00000
#> 6   628.3400            19.42940            19.42940            19.42940
#> 7   754.0081            41.42912            41.42912            41.42912
#> 8   879.6761            63.14696            63.14696            63.14696
#> 9  1005.3441            84.75038            84.75038            84.75038
#> 10 1131.0121           106.35659           106.35659           106.35659
#> 11 1256.6801           133.68948           133.68948           133.68948
#> 12 1382.3481           160.34701           160.34701           160.34701
#> 13 1508.0161           183.58582           183.58582           183.58582
#> 14 1633.6841           210.42614           210.42614           210.42614
#> 15 1759.3521           237.26646           237.26646           237.26646
#> 16 1885.0201           263.80119           263.80119           263.80119
#> 17 2010.6881           290.33592           290.33592           290.33592
#> 18 2136.3562           316.55947           316.55947           316.55947
#> 19 2262.0242           342.78303           342.78303           342.78303
#> 20 2387.6922           367.34328           367.34328           367.34328
#> 21 2513.3602           391.90353           391.90353           391.90353
#> 22 2639.0282           420.94856           420.94856           420.94856
#> 23 2764.6962           449.99360           449.99360           449.99360
#> 24 2890.3642           477.54696           477.54696           477.54696
#> 25 3016.0322           505.10032           505.10032           505.10032
#> 26 3141.7002           533.72813           533.72813           533.72813
#> 27 3267.3682           562.35595           562.35595           562.35595
#> 28 3393.0362           586.35525           586.35525           586.35525
#> 29 3518.7043           610.35456           610.35456           610.35456
#> 30 3644.3723           635.87484           635.87484           635.87484
#> 31 3770.0403           661.39512           661.39512           661.39512
#> 32 3895.7083           688.35683           688.35683           688.35683
#> 33 4021.3763           715.31855           715.31855           715.31855
#> 34 4147.0443           807.06262           807.06262           807.06262
#> 35 4272.7123           898.80670           898.80670           898.80670
#> 36 4398.3803           937.48550           937.48550           937.48550
#> 37 4524.0483           976.16430           976.16430           976.16430
#>          watershed
#> 1  Mokelumne River
#> 2  Mokelumne River
#> 3  Mokelumne River
#> 4  Mokelumne River
#> 5  Mokelumne River
#> 6  Mokelumne River
#> 7  Mokelumne River
#> 8  Mokelumne River
#> 9  Mokelumne River
#> 10 Mokelumne River
#> 11 Mokelumne River
#> 12 Mokelumne River
#> 13 Mokelumne River
#> 14 Mokelumne River
#> 15 Mokelumne River
#> 16 Mokelumne River
#> 17 Mokelumne River
#> 18 Mokelumne River
#> 19 Mokelumne River
#> 20 Mokelumne River
#> 21 Mokelumne River
#> 22 Mokelumne River
#> 23 Mokelumne River
#> 24 Mokelumne River
#> 25 Mokelumne River
#> 26 Mokelumne River
#> 27 Mokelumne River
#> 28 Mokelumne River
#> 29 Mokelumne River
#> 30 Mokelumne River
#> 31 Mokelumne River
#> 32 Mokelumne River
#> 33 Mokelumne River
#> 34 Mokelumne River
#> 35 Mokelumne River
#> 36 Mokelumne River
#> 37 Mokelumne River