Paynes Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

paynes_creek_floodplain

Format

dataframe with 30 rows and 4 variables:

flow_cfs

integer flow value in cubic feet per second

FR_floodplain_acres

fall run floodplain acreage

ST_floodplain_acres

steelhead floodplain acreage

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Fall Run Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Deer Creek based onhydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.22 percent of DeerCreek values based on the ratio of mean flow from December to June between this watershed and Deer Creek. Of the entire mapped 11.81 mile rearing extent in this watershed, 2.04 miles were classified as low gradient and 9.77 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Steelhead Modeling

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Deer Creek based onhydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.22 percent of DeerCreek values based on the ratio of mean flow from December to June between this watershed and Deer Creek. Of the entire mapped 27.81 mile rearing extent in this watershed, 2.04 miles were classified as low gradient and 25.77 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.

Examples

paynes_creek_floodplain

#>     flow_cfs FR_floodplain_acres ST_floodplain_acres    watershed
#> 1   15.57264           0.0000000           0.0000000 Paynes Creek
#> 2   37.81927           0.0000000           0.0000000 Paynes Creek
#> 3   60.06589           0.0000000           0.0000000 Paynes Creek
#> 4   82.31252           0.0000000           0.0000000 Paynes Creek
#> 5  104.55915           0.0000000           0.0000000 Paynes Creek
#> 6  126.80577           0.0000000           0.0000000 Paynes Creek
#> 7  127.02824           0.0000000           0.0000000 Paynes Creek
#> 8  149.05240           0.0000000           0.0000000 Paynes Creek
#> 9  171.29903           0.0000000           0.0000000 Paynes Creek
#> 10 193.54565           0.1404349           0.2149200 Paynes Creek
#> 11 215.79228           0.4124257           0.6311714 Paynes Creek
#> 12 238.03891           0.6405593           0.9803044 Paynes Creek
#> 13 260.28554           0.8859152           1.3557942 Paynes Creek
#> 14 282.53216           1.1028507           1.6877896 Paynes Creek
#> 15 304.77879           1.3341667           2.0417930 Paynes Creek
#> 16 327.02542           1.5525380           2.3759859 Paynes Creek
#> 17 349.27204           1.7637846           2.6992752 Paynes Creek
#> 18 371.51867           1.9451183           2.9767862 Paynes Creek
#> 19 393.76530           2.1186918           3.2424210 Paynes Creek
#> 20 416.01192           2.2819633           3.4922898 Paynes Creek
#> 21 438.25855           2.4260956           3.7128681 Paynes Creek
#> 22 460.50518           2.5778964           3.9451823 Paynes Creek
#> 23 482.75181           2.6961709           4.1261881 Paynes Creek
#> 24 504.99843           2.8151960           4.3083429 Paynes Creek
#> 25 527.24506           2.9183313           4.4661798 Paynes Creek
#> 26 549.49169           3.0312025           4.6389167 Paynes Creek
#> 27 571.73831           3.1402729           4.8058367 Paynes Creek
#> 28 593.98494           3.2469339           4.9690694 Paynes Creek
#> 29 616.23157           3.3487275           5.1248531 Paynes Creek
#> 30 638.47819           3.4305120           5.2500151 Paynes Creek