Paynes Creek Flow to Floodplain Habitat Area Relationship

A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.

paynes_creek_floodplain

Format

dataframe with 22 rows and 3 variables:

flow_cfs

integer flow value in cubic feet per second

floodplain_acres

fall run floodplain acreage (use for Spring Run and Steelhead when required)

watershed

name of watershed

Source

Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model

Details

HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Paynes Creek scaling factor of 22% was calculated as the average ratio of Paynes Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 2.0 mile low gradient and 9.8 miles of high gradient mapped rearing extent in paynes Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 29.6 acres) and floodplain inundation area.

Examples

paynes_creek_floodplain

#> # A tibble: 22 x 3
#>    flow_cfs floodplain_acres    watershed
#>       <dbl>            <dbl>        <chr>
#>  1   171.30             0.00 Paynes Creek
#>  2   193.55             0.14 Paynes Creek
#>  3   215.79             0.41 Paynes Creek
#>  4   238.04             0.64 Paynes Creek
#>  5   260.29             0.89 Paynes Creek
#>  6   282.53             1.10 Paynes Creek
#>  7   304.78             1.33 Paynes Creek
#>  8   327.03             1.55 Paynes Creek
#>  9   349.27             1.76 Paynes Creek
#> 10   371.52             1.95 Paynes Creek
#> # ... with 12 more rows