Set Instream Habitat Area

This function returns an estimated instream habitat area based on watershed, species, life stage and flow.

set_instream_habitat(watershed, species, life_stage, flow)

Arguments

watershed
one of the watersheds defined for the SIT model

species
one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)

life_stage
life stage of fish, one of 'juv', 'adult' or 'fry'

flow
value used to determine habitat area

Value

habitat area in square meters

Examples

# Fry rearing habitat value in square meters for Fall Run in the Merced River at 425 cfs.
set_instream_habitat('Merced River', 'fr', 'fry', 425) # habitat modeling exists

#> [1] 439269.2

# Juvenile rearing habitat value in square meters for Fall Run in Elder Creek at 300 cfs.
set_instream_habitat('Elder Creek', 'fr', 'juv', 300) # no habitat modeling exists, composite used

#> [1] 37462.14