Set Spawning Habitat Area

This function returns an estimated spawning area based on watershed, species, and flow.

set_spawning_habitat(watershed, species, flow)

Arguments

watershed
one of the watersheds defined for the SIT model

species
one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)

flow
a flow value in cubic feet per second used to determine habitat area

Value

habitat area in square meters

Details

The function relies on a dataframe called modeling_exist that contains data on whether the species is present in a watershed and whether habitat modeling exists. If a model for the watershed does exist, the function looks up the flow to weighted usable area (WUA) relationship (e.g. battle_creek_instream) and selects the correct WUA for the given flow and species. This WUA is then multiplied by the watershed's typical spawning habitat extent length (stored in watershed_lengths), to return an estimate of suitable spawning habitat within the watershed. When additional species modeling is not available, the fall run WUA values are used (lengths are modified if the habitat extent varies across species).

Regional Approximation: When a watershed has no associated flow to WUA reltionship, a regional approximation is made. First, the mean WUA at the given flow vale from a set of similar modeled watersheds nearby is calculated. Then the mean WUA is multiplied by the river length of the watershed of interest.

Below are the regions (defined by the downstream watershed) that contain watersheds with unmodeled spawning relationships. The modeled watersheds used to approximate spawning area for the unmodeled watersheds are marked with an asterisk.

Upper-mid Sacramento River

Battle Creek*

Bear Creek

Big Chico Creek

Butte Creek*

Clear Creek*

Cottonwood Creek*

Cow Creek

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

South Delta

Calaveras River*

Cosumnes River

Mokelumne River*

Examples

# determine the spawning habitat area at Cottonwood Creek for Fall Run Chinook at flow 1567
set_spawning_habitat("Cottonwood Creek", "fr", 1567)

#> [1] 6003.183

# determine spawning habitat for a watershed with no modeling, uses region approximation
set_spawning_habitat("Antelope Creek", "fr", 1597)

#> [1] 14664.56