Set spawning habitat area based on watershed, species, life stage and flow.

Set spawning habitat area based on watershed, species, life stage and flow.

set_spawning_habitat(watershed, species, flow)

Arguments

watershed
one of the watersheds defined for the SIT model

species
one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)

flow
a flow value in cubic feet per second used to determine habitat area

Value

habitat area in square meters

Modeling Existance

The set_spawning_habitat function relies on a dataframe exported by the package called cvpiaHabitat::modeling_exist. This dataframe allows the set_spawning_habitat function to determine whether a given watershed has either no existance of the species, has existance but no model, or whether a model exists. Changes to the dataframe will affect whether an attempt is made to look for the relationship between flow and wua or a region approximation is made to determine habitat area. The process for updating this dataframe requires not only altering the dataframe, but adding a relationship to package in the form of a dataframe using devtools::use_data(watershed_name_instream). For more details see here

Region Approximation

When a watershed has no associated flow to wua reltionship a region approximation is made. Simply put, this approach determines a set of related watersheds (defined as a region) for which a model does exists. It then applies set_spawning_habitat for each, takes a mean of them and returns the appropriate area from this calculation.

Examples

# determine the spawning habitat area at Cottonwood Creek for Fall Run Chinook at flow 1567
set_spawning_habitat("Cottonwood Creek", "fr", 1567)

#> [1] 6003.183

# determine spawning habitat for a watershed with no modeling, uses region approximation
set_spawning_habitat("Antelope Creek", "fr", 1597)

#> [1] 14664.56