Mainstem Sacramento River Spawning Habitat

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second.

upper_sac_ACID_boards_in

upper_sac_ACID_boards_out

Format

dataframe with 30 rows and 11 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_WUA

fall run spawning WUA in square feet per 1000 feet

LFR_spawn_WUA

late-fall run fry spawning WUA in square feet per 1000 feet

ST_spawn_WUA

steelhead spawning WUA in square feet per 1000 feet

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

The spawning WUA were derived from a FWS River2D model. The study is broken into several segments: Keswick to A.C.I.D. (3.5 mi), A.C.I.D. to Cow Creek (18.5 mi), Cow Creek to Battle Creek (8.5 mi), Battle Creek to Red Bluff (22.5 mi), and Red Bluff to Deer Creek (23.5 mi). The Upper Sacramento fall run spawning WUA values include the spawning that occurs in the Upper and Upper-mid Sacramento River (Keswick to Deer Creek). The Late-Fall Run, Winter Run, and Steelhead spawn from Keswick to Battle Creek. The A.C.I.D. boards are in April 1st - October 31st

Species 	 Migration 	 Peak Migration 	 Spawning 	 Peak Spawning 	 Juvenile Emergence 	 Juvenile Rearing

Late-Fall Run 	 Oct-Apr 	 Dec 	 Jan-Apr 	 Feb-Mar 	 Apr-Jun 	 7-13 mths

Winter Run 	 Dec-Jul 	 Mar 	 Apr-Aug 	 May-Jun 	 Jul-Oct 	 5-10 mths

Spring Run 	 Mar-Sep 	 May-Jun 	 Aug-Oct 	 Mid-Sep 	 Nov-Mar 	 3-15 mths

Fall Run 	 Jun-Dec 	 Sep-Oct 	 Sep-Dec 	 Oct-Nov 	 Dec-Mar 	 1-7 mths

Generalised life history Yoshinyama et al. 1998

Examples

upper_sac_ACID_boards_in

#> # A tibble: 30 x 6
#>    flow_cfs FR_spawn_WUA LFR_spawn_WUA ST_spawn_WUA WR_spawn_WUA
#>       <dbl>        <dbl>         <dbl>        <dbl>        <dbl>
#>  1     3250     8553.719      7945.461     1998.007     5321.020
#>  2     3500     8756.922      8204.024     2034.904     5627.105
#>  3     3750     8939.986      8329.719     2041.319     5880.905
#>  4     4000     9069.236      8394.523     2025.323     6111.320
#>  5     4250     9082.162      8429.446     2009.004     6350.286
#>  6     4500     9192.716      8422.957     2004.800     6580.079
#>  7     4750     9311.547      8375.137     1998.802     6795.659
#>  8     5000     9282.365      8309.960     1997.398     6981.632
#>  9     5250     9299.413      8228.626     2002.956     7126.329
#> 10     5500     9254.085      8119.629     2007.923     7278.099
#> # ... with 20 more rows, and 1 more variables: watershed <chr>