Upper-mid Sacramento River Region Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

upper_mid_sac_region_instream

Format

dataframe with 14 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Sadie Gill

Details

Some of the Upper Sacramento Rivers' instream rearing habitat have not been modeled. The rearing WUA values were estimated using the mean WUA at each flow from Battle Creek, Butte Creek, Clear Creek, and Cow Creek. The spawning WUA values were estimated using the mean WUA at each flow from Battle Creek, Butte Creek, and Clear Creek.

Spawning Not Modeled:

Antelope Creek

Bear Creek

Big Chico Creek

Cow Creek

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

Rearing Not Modeled:

Antelope Creek

Bear Creek

Big Chico Creek

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

Examples

upper_mid_sac_region_instream

#> # A tibble: 23 x 5
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua
#>       <int>        <dbl>      <dbl>      <dbl>
#>  1       50     1680.581  13123.573  11114.864
#>  2       75     2010.154  12408.408  12034.163
#>  3      100     2187.526  11471.737  12448.544
#>  4      125     2279.997  10641.957  12463.118
#>  5      150     2349.041  10026.154  12230.094
#>  6      175     2416.404   9476.974  11802.526
#>  7      200     2444.635   9078.136  11316.242
#>  8      225     2431.503   8791.203  10786.358
#>  9      250     2399.343   8571.605  10229.627
#> 10      275     2346.695   8414.557   9722.572
#> # ... with 13 more rows, and 1 more variables: watershed <chr>