Upper-mid Sacramento River Region Instream Flow to Habitat Area Relationship

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second

upper_mid_sac_region_instream

Format

dataframe with 14 rows and 5 variables:

flow_cfs

integer flow value in cubic feet per second

FR_spawn_wua

spawning WUA in square feet per 1000 feet

FR_fry_wua

fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua

juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Sadie Gill

Details

Some of the Upper Sacramento Rivers' instream rearing habitat have not been modeled. The rearing WUA values were estimated using the mean WUA at each flow from Battle Creek, Butte Creek, Clear Creek, and Cow Creek. The spawning WUA values were estimated using the mean WUA at each flow from Battle Creek, Butte Creek, and Clear Creek.

Spawning Not Modeled:

Antelope Creek

Bear Creek

Big Chico Creek

Cow Creek

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

Rearing Not Modeled:

Antelope Creek

Bear Creek

Big Chico Creek

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

Examples

upper_mid_sac_region_instream

#> # A tibble: 23 x 5
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua watershed                        
#>       <int>        <dbl>      <dbl>      <dbl> <chr>                            
#>  1       50        1681.     13124.     11115. Upper-mid Sacramento River Region
#>  2       75        2010.     12408.     12034. Upper-mid Sacramento River Region
#>  3      100        2188.     11472.     12449. Upper-mid Sacramento River Region
#>  4      125        2280.     10642.     12463. Upper-mid Sacramento River Region
#>  5      150        2349.     10026.     12230. Upper-mid Sacramento River Region
#>  6      175        2416.      9477.     11803. Upper-mid Sacramento River Region
#>  7      200        2445.      9078.     11316. Upper-mid Sacramento River Region
#>  8      225        2432.      8791.     10786. Upper-mid Sacramento River Region
#>  9      250        2399.      8572.     10230. Upper-mid Sacramento River Region
#> 10      275        2347.      8415.      9723. Upper-mid Sacramento River Region
#> # ... with 13 more rows