Upper Sacramento River Instream Flow to Habitat Area Relationship (A.C.I.D. Boards In)

A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second.

upper_sac_ACID_boards_in

Format

dataframe with 30 rows and 11 variables:

flow_cfs

integer flow value in cubic feet per second

FR_fry_WUA

fall run fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_WUA

fall run juvenile WUA in square feet per 1000 feet

FR_spawn_WUA

fall run spawning WUA in square feet per 1000 feet

LFR_fry_WUA

late-fall run fry (up to 50 mm) WUA in square feet per 1000 feet

LFR_juv_WUA

late-fall run fry juvenile WUA in square feet per 1000 feet

LFR_spawn_WUA

late-fall run fry spawning WUA in square feet per 1000 feet

ST_spawn_WUA

steelhead spawning WUA in square feet per 1000 feet

WR_fry_WUA

winter run fry (up to 50 mm) WUA in square feet per 1000 feet

WR_juv_WUA

winter run juvenile WUA in square feet per 1000 feet

watershed

name of watershed

Source

Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model

Details

The spawning WUA were derived from a FWS River2D model. The study is broken into several segments: Keswick to A.C.I.D. (3.5 mi), A.C.I.D. to Cow Creek (18.5 mi), Cow Creek to Battle Creek (8.5 mi), Battle Creek to Red Bluff (22.5 mi), and Red Bluff to Deer Creek (23.5 mi). The Upper Sacramento fall run spawning WUA values include the spawning that occurs in the Upper and Upper-mid Sacramento River (Keswick to Deer Creek). The Late-Fall Run, Winter Run, and Steelhead spawn from Keswick to Battle Creek. The rearing WUA values come from two models, River2D (Keswick to Battle Creek) and HEC-RAS 1D (Battle Creek to Red Bluff). The A.C.I.D. boards are in April 1st - October 31st

Species 	 Migration 	 Peak Migration 	 Spawning 	 Peak Spawning 	 Juvenile Emergence 	 Juvenile Rearing

Late-Fall Run 	 Oct-Apr 	 Dec 	 Jan-Apr 	 Feb-Mar 	 Apr-Jun 	 7-13 mths

Winter Run 	 Dec-Jul 	 Mar 	 Apr-Aug 	 May-Jun 	 Jul-Oct 	 5-10 mths

Spring Run 	 Mar-Sep 	 May-Jun 	 Aug-Oct 	 Mid-Sep 	 Nov-Mar 	 3-15 mths

Fall Run 	 Jun-Dec 	 Sep-Oct 	 Sep-Dec 	 Oct-Nov 	 Dec-Mar 	 1-7 mths

Generalised life history Yoshinyama et al. 1998

Examples

upper_sac_ACID_boards_in

#> # A tibble: 30 x 12
#>    flow_cfs FR_fry_WUA FR_juv_WUA FR_spawn_WUA LFR_fry_WUA LFR_juv_WUA
#>       <dbl>      <dbl>      <dbl>        <dbl>       <dbl>       <dbl>
#>  1     3250  12165.797   4582.240     8553.719   11294.604    4150.801
#>  2     3500  11867.933   4499.783     8756.922   11066.133    4082.004
#>  3     3750  11920.436   4419.088     8939.986   11199.441    4013.369
#>  4     4000  11726.850   4357.855     9069.236   11074.919    3962.202
#>  5     4250  11432.613   4346.044     9082.162   10809.985    3952.086
#>  6     4500  11055.769   4303.682     9192.716   10482.806    3917.629
#>  7     4750  10605.775   4226.807     9311.547   10102.670    3851.167
#>  8     5000  10152.062   4241.114     9282.365    9692.486    3773.503
#>  9     5250   9765.909   4057.402     9299.413    9304.173    3702.602
#> 10     5500   9398.373   3981.470     9254.085    8897.696    3637.407
#> # ... with 20 more rows, and 6 more variables: LFR_spawn_WUA <dbl>,
#> #   ST_spawn_WUA <dbl>, WR_fry_WUA <dbl>, WR_juv_WUA <dbl>, WR_spawn_WUA <dbl>,
#> #   watershed <chr>