Habitat Extent Lengths

A dataset containing the length of rearing and spawning extent within each CVPIA watershed for different species.

watershed_lengths

Format

dataframe with 57 rows and 7 variables

order

integer value representing watershed order in SIT model

watershed

name of watershed

lifestage

habitat type by lifestage, 'spawing' or 'rearing'

miles

length in miles

feet

length in feet

source

Expert who delineated habitat extents

species

species of habitat extent

Source

FlowWest, add url to shp file when mapping is complete

Details

Information comes from expert outreach undertaken in 2017.

Examples

watershed_lengths

#> # A tibble: 56 x 7
#>    order              watershed lifestage     miles      feet          source
#>    <int>                  <chr>     <chr>     <dbl>     <dbl>           <chr>
#>  1     0            North Delta   rearing 45.654083 241053.56      Mike Urkov
#>  2     0            South Delta   rearing 69.617216 367578.90      Mike Urkov
#>  3     1 Upper Sacramento River  spawning 59.350116 313368.61       Jim Smith
#>  4     1 Upper Sacramento River   rearing 59.185680 312500.39       Jim Smith
#>  5     2         Antelope Creek  spawning 23.685584 125059.88    Matt Johnson
#>  6     2         Antelope Creek   rearing 31.619077 166948.73    Matt Johnson
#>  7     3           Battle Creek  spawning  3.966914  20945.31       Jim Smith
#>  8     3           Battle Creek   rearing  5.871706  31002.61       Jim Smith
#>  9     4             Bear Creek  spawning 17.103907  90308.63 Tricia Bratcher
#> 10     4             Bear Creek   rearing 17.296223  91324.06 Tricia Bratcher
#> # ... with 46 more rows, and 1 more variables: species <chr>