Flow Thresholds for Number of Weeks Inundated

Estimated relationship between duration of inundation and monthly mean flow

weeks_inundated

Format

dataframe with rows and 3 variables

watershed

name of CVPIA watershed

weeks_inundated

integer value between 0 and 4 to represent weeks of floodplain inundation

flow_threshold

flow threshold in cubic feet per second associated with number of weeks inundated

Source

Sadie Gill sgill@flowwest.com

Details

Visual inspection of measured flow and professional judgement were used to define relationships between number of days inundated and the mean monthly flow.

Relationship Modeled:

Sacramento River

Big Chico Creek

Butte Creek

Cottonwood Creek

Deer Creek

Elder Creek

Sutter Bypass

Bear River

Feather River

Yuba River

Yolo Bypass

American River

Calaveras River

Merced River

Stanislaus River

Tuolumne River

San Joaquin River

All other watersheds are assumed to have a two week inundation.

Examples

weeks_inundated

#>            watershed weeks_inundated flow_threshold
#> 1         Yuba River               0              0
#> 2         Yuba River               1             NA
#> 3         Yuba River               2             NA
#> 4         Yuba River               3             NA
#> 5         Yuba River               4            800
#> 6    Big Chico Creek               0              0
#> 7    Big Chico Creek               1             NA
#> 8    Big Chico Creek               2            265
#> 9    Big Chico Creek               3             NA
#> 10   Big Chico Creek               4            900
#> 11       Butte Creek               0              0
#> 12       Butte Creek               1             NA
#> 13       Butte Creek               2            310
#> 14       Butte Creek               3             NA
#> 15       Butte Creek               4            500
#> 16  Cottonwood Creek               0              0
#> 17  Cottonwood Creek               1             NA
#> 18  Cottonwood Creek               2             NA
#> 19  Cottonwood Creek               3             NA
#> 20  Cottonwood Creek               4            400
#> 21        Deer Creek               0              0
#> 22        Deer Creek               1             NA
#> 23        Deer Creek               2            370
#> 24        Deer Creek               3             NA
#> 25        Deer Creek               4            750
#> 26       Elder Creek               0              0
#> 27       Elder Creek               1             NA
#> 28       Elder Creek               2            130
#> 29       Elder Creek               3             NA
#> 30       Elder Creek               4            250
#> 31        Bear River               0              0
#> 32        Bear River               1             NA
#> 33        Bear River               2            760
#> 34        Bear River               3             NA
#> 35        Bear River               4           1000
#> 36     Feather River               0              0
#> 37     Feather River               1             NA
#> 38     Feather River               2             NA
#> 39     Feather River               3             NA
#> 40     Feather River               4           4000
#> 41    American River               0              0
#> 42    American River               1             NA
#> 43    American River               2             NA
#> 44    American River               3             NA
#> 45    American River               4           4200
#> 46   Calaveras River               0              0
#> 47   Calaveras River               1             NA
#> 48   Calaveras River               2             NA
#> 49   Calaveras River               3            210
#> 50   Calaveras River               4             NA
#> 51      Merced River               0              0
#> 52      Merced River               1             NA
#> 53      Merced River               2             NA
#> 54      Merced River               3             NA
#> 55      Merced River               4            500
#> 56  Stanislaus River               0              0
#> 57  Stanislaus River               1             NA
#> 58  Stanislaus River               2             NA
#> 59  Stanislaus River               3             NA
#> 60  Stanislaus River               4            500
#> 61    Tuolumne River               0              0
#> 62    Tuolumne River               1             NA
#> 63    Tuolumne River               2             NA
#> 64    Tuolumne River               3             NA
#> 65    Tuolumne River               4           1000
#> 66 San Joaquin River               0              0
#> 67 San Joaquin River               1             NA
#> 68 San Joaquin River               2             NA
#> 69 San Joaquin River               3             NA
#> 70 San Joaquin River               4           1520